% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adisit.R
\name{Adisit}
\alias{Adisit}
\alias{cdo_adipot}
\alias{cdo_adisit}
\title{Potential temperature to in-situ temperature and vice versa}
\usage{
cdo_adipot(input_file, output_file = NULL)

cdo_adisit(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
DESCRIPTION
}
\details{
\preformatted{ adisit  Potential temperature to in-situ temperature
         This is a special operator for the post processing of the ocean and sea ice model output.
         It converts potential temperature adiabatically to in-situ temperature to(t, s, p).
         Required input fields are sea water potential temperature (name=tho; code=2) and sea water salinity (name=sao; code=5).
         Pressure is calculated from the level information or can be specified by the optional parameter.
         Output fields are sea water temperature (name=to; code=20) and sea water salinity (name=s; code=5).
 adipot  In-situ temperature to potential temperature
         This is a special operator for the post processing of the ocean and sea ice
         model outpu.  It converts in-situ temperature to potential temperature tho(to,
         s, p).  Required input fields are sea water in-situ temperature (name=t; code=2) 
         and sea water salinity (name=sao,s; code=5).  Pressure is calculated
         from the level information or can be specified by the optional parameter.
         Output fields are sea water temperature (name=tho; code=2) and sea water
         salinity (name=s; code=5).
}
}
