% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterburner.R
\name{Afterburner}
\alias{Afterburner}
\alias{cdo_after}
\title{ECHAM standard post processor}
\usage{
cdo_after(infiles, outfile = NULL)
}
\arguments{
\item{infiles}{input files.}

\item{outfile}{output file.}
}
\description{
The \"afterburner\" is the standard post processor for ECHAM data which
provides the following operations:
}
\details{
\preformatted{ gmtxyz    GMT xyz format
           The operator exports the first field to the GMT xyz ASCII format.
           The output can be used to create contour plots with the GMT module pscontour.
 gmtcells  GMT multiple segment format
           The operator exports the first field to the GMT multiple segment ASCII format.
           The output can be used to create shaded gridfill plots with the GMT module psxy.
}
}
