% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithc.R
\name{Arithc}
\alias{Arithc}
\alias{cdo_addc}
\alias{cdo_divc}
\alias{cdo_mulc}
\alias{cdo_subc}
\title{Arithmetic with a constant}
\usage{
cdo_addc(input_file, output_file = NULL)

cdo_divc(input_file, output_file = NULL)

cdo_mulc(input_file, output_file = NULL)

cdo_subc(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module performs simple arithmetic with all field elements of a dataset and
a constant. The fields in outfile inherit the meta data from infile.
}
\details{
\preformatted{ addc  Add a constant
       o(t,x) = i(t,x) + c
 subc  Subtract a constant
       o(t,x) = i(t,x) - c
 mulc  Multiply with a constant
       o(t,x) = i(t,x) * c
 divc  Divide by a constant
       o(t,x) = i(t,x) / c
 minc  Minimum of a field and a constant
       o(t,x) = min(i(t,x), c)
 maxc  Maximum of a field and a constant
       o(t,x) = max(i(t,x), c)
}
}
