% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithdays.R
\name{Arithdays}
\alias{Arithdays}
\alias{cdo_divdpm}
\alias{cdo_divdpy}
\alias{cdo_muldpm}
\alias{cdo_muldpy}
\title{Arithmetic with days}
\usage{
cdo_divdpm(infile, outfile = NULL)

cdo_divdpy(infile, outfile = NULL)

cdo_muldpm(infile, outfile = NULL)

cdo_muldpy(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module multiplies or divides each timestep of a dataset with the
corresponding days per month or days per year. The result of these functions
depends on the used calendar of the input data.
}
\details{
\preformatted{ muldpm  Multiply with days per month
         o(t,x) = i(t,x) * days_per_month
 divdpm  Divide by days per month
         o(t,x) = i(t,x) / days_per_month
 muldpy  Multiply with days per year
         o(t,x) = i(t,x) * days_per_year
 divdpy  Divide by days per year
         o(t,x) = i(t,x) / days_per_year
}
}
