% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-cdo_run.R
\name{cdo_add_output}
\alias{cdo_add_output}
\title{Add output to operators}
\usage{
cdo_add_output(cdo_operator, output_files)
}
\arguments{
\item{cdo_operator}{a cdo_operator, which is the result of any \verb{cdo_*} functions
that has at least one output file. `}

\item{output_files}{filenames}
}
\description{
Easily add an output file to an operator.
}
