% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-cdo_run.R
\name{cdo_run}
\alias{cdo_run}
\title{Run a cdo operation.}
\usage{
cdo_run(operation)
}
\arguments{
\item{operation}{a cdo_operator, which is the result of any \verb{cdo_*} functions
that has at least one output file. `}
}
\description{
The various cdo functions only build command and this function runs
them and returns the output. The exceptions are operator with no output
files (such as \code{\link[=cdo_info]{cdo_info()}}), in which case, the commands are executed
inmmediately.
}
