% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{Change}
\alias{Change}
\alias{cdo_chcode}
\alias{cdo_chlevel}
\alias{cdo_chlevelc}
\alias{cdo_chlevelv}
\alias{cdo_chname}
\alias{cdo_chparam}
\alias{cdo_chunit}
\title{Change field header}
\usage{
cdo_chcode(input_file, output_file = NULL)

cdo_chlevel(input_file, output_file = NULL)

cdo_chlevelc(input_file, output_file = NULL)

cdo_chlevelv(input_file, output_file = NULL)

cdo_chname(input_file, output_file = NULL)

cdo_chparam(input_file, output_file = NULL)

cdo_chunit(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module reads fields from infile, changes some header values and writes the
results to outfile. The kind of changes depends on the chosen operator.
}
\details{
\preformatted{ chcode    Change code number
           Changes some user given code numbers to new user given values.
 chparam   Change parameter identifier
           Changes some user given parameter identifiers to new user given values.
 chname    Change variable or coordinate name
           Changes some user given variable or coordinate names to new user given names.
 chunit    Change variable unit
           Changes some user given variable units to new user given units.
 chlevel   Change level
           Changes some user given levels to new user given values.
 chlevelc  Change level of one code
           Changes one level of a user given code number.
 chlevelv  Change level of one variable
           Changes one level of a user given variable name.
}
}
