% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmor.R
\name{CMOR}
\alias{CMOR}
\alias{cdo_cmor}
\title{Climate Model Output Rewriting to produce CMIP-compliant data}
\usage{
cdo_cmor(input_file)
}
\arguments{
\item{input_file}{input file.}
}
\description{
DESCRIPTION
}
\details{
\preformatted{ uv2vr_cfd  U and V wind to relative vorticity
            Computes relative vorticity for a latitude-longitude grid using centered finite differences.
            The grid need not be global and missing values are allowed.
 uv2dv_cfd  U and V wind to divergence
            Computes divergence for a latitude-longitude grid using centered finite differences.
            The grid need not be global and missing values are allowed.
}
}
