% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmorlite.R
\name{CMORlite}
\alias{CMORlite}
\alias{cdo_cmorlite}
\title{CMOR lite}
\usage{
cdo_cmorlite(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
The CMOR (Climate Model Output Rewriter) library comprises a set of functions,
that can be used to produce CF-compliant NetCDF files that fulfill the
requirements of many of the climate community's standard model experiments.
These experiments are collectively referred to as MIP's. Much of the metadata
written to the output files is defined in MIP-specific tables, typically made
available from each MIP's web site.
}
\details{
\preformatted{ histcount  Histogram count
            Number of elements in the bin range.
 histsum    Histogram sum
            Sum of elements in the bin range.
 histmean   Histogram mean
            Mean of elements in the bin range.
 histfreq   Histogram frequency
            Relative frequency of elements in the bin range.
}
}
