% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond.R
\name{Cond}
\alias{Cond}
\alias{cdo_ifnotthen}
\alias{cdo_ifthen}
\title{Conditional select one field}
\usage{
cdo_ifnotthen(input_file_1, input_file_2, output_file = NULL)

cdo_ifthen(input_file_1, input_file_2, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2}{input files.}

\item{output_file}{output file.}
}
\description{
This module selects field elements from infile2 with respect to infile1 and
writes them to outfile. The fields in infile1 are handled as a mask. A value
not equal to zero is treated as \"true\ zero is treated as \"false\". The
number of fields in infile1 has either to be the same as in infile2 or the same
as in one timestep of infile2 or only one. The fields in outfile inherit the
meta data from infile2.
}
\details{
\preformatted{ ifthen     If then
                     / i_2(t,x) if i_1([t,]x) NE 0  AND  i_1([t,]x) NE miss
            o(t,x) =
                     \\\\ miss     if i_1([t,]x) EQ 0  OR   i_1([t,]x) EQ miss
 ifnotthen  If not then
                     / i_2(t,x) if i_1([t,]x) EQ 0  AND  i_1([t,]x) NE miss
            o(t,x) = 
                     \\\\ miss     if i_1([t,]x) NE 0  OR   i_1([t,]x) EQ miss
}
}
