% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condc.R
\name{Condc}
\alias{Condc}
\alias{cdo_ifnotthenc}
\alias{cdo_ifthenc}
\title{Conditional select a constant}
\usage{
cdo_ifnotthenc(infile, outfile = NULL)

cdo_ifthenc(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module creates fields with a constant value or missing value. The fields
in infile are handled as a mask. A value not equal to zero is treated as
\"true\ zero is treated as \"false\".
}
\details{
\preformatted{ ifthenc     If then constant
                      / c      if i(t,x) NE 0  AND  i(t,x) NE miss
             o(t,x) =
                      \\\\ miss   if i(t,x) EQ 0  OR   i(t,x) EQ miss
 ifnotthenc  If not then constant
                      / c      if i(t,x) EQ 0  AND  i(t,x) NE miss
             o(t,x) =
                      \\\\ miss   if i(t,x) NE 0  OR   i(t,x) EQ miss
}
}
