% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecstat.R
\name{Consecstat}
\alias{Consecstat}
\alias{cdo_consecsum}
\alias{cdo_consects}
\title{Consecute timestep periods}
\usage{
cdo_consecsum(infile, outfile = NULL)

cdo_consects(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes periods over all timesteps in infile where a certain
property is valid. The property can be chosen by creating a mask from the
original data, which is the expected input format for operators of this module.
Depending on the operator full information about each period or just its length
and ending date are computed.
}
\details{
\preformatted{ consecsum  Consecutive Sum
            This operator computes periods of consecutive timesteps similar to a
            runsum, but periods are finished, when the mask value is 0. That way
            multiple periods can be found. Timesteps from the input are preserved. Missing
            values are handled like 0, i.e. finish periods of consecutive timesteps.
 consects   Consecutive Timesteps
            In contrast to the operator above consects only computes the length of each
            period together with its last timestep. To be able to perform statistical
            analysis like min, max or mean, everything else is set to missing value.
}
}
