% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{Copy}
\alias{Copy}
\alias{cdo_cat}
\alias{cdo_copy}
\title{Copy datasets}
\usage{
cdo_cat(input_files, output_file = NULL)

cdo_copy(input_files, output_file = NULL)
}
\arguments{
\item{input_files}{input files.}

\item{output_file}{output file.}
}
\description{
This module contains operators to copy or concatenate datasets. infiles is an
arbitrary number of input files. All input files need to have the same
structure with the same variables on different timesteps.
}
\details{
\preformatted{ copy  Copy datasets
       Copies all input datasets to outfile. 
 cat   Concatenate datasets
       Concatenates all input datasets and appends the result to the end 
       of outfile. If outfile does not exist it will be created.
}
}
