% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daystat.R
\name{Daystat}
\alias{Daystat}
\alias{cdo_dayavg}
\alias{cdo_daymax}
\alias{cdo_daymean}
\alias{cdo_daymin}
\alias{cdo_dayrange}
\alias{cdo_daystd}
\alias{cdo_daystd1}
\alias{cdo_daysum}
\alias{cdo_dayvar}
\alias{cdo_dayvar1}
\title{Daily statistical values}
\usage{
cdo_dayavg(input_file, output_file = NULL)

cdo_daymax(input_file, output_file = NULL)

cdo_daymean(input_file, output_file = NULL)

cdo_daymin(input_file, output_file = NULL)

cdo_dayrange(input_file, output_file = NULL)

cdo_daystd(input_file, output_file = NULL)

cdo_daystd1(input_file, output_file = NULL)

cdo_daysum(input_file, output_file = NULL)

cdo_dayvar(input_file, output_file = NULL)

cdo_dayvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values over timesteps of the same day.
Depending on the chosen operator the minimum, maximum, range, sum, average,
variance or standard deviation of timesteps of the same day is written to
outfile. The time of outfile is determined by the time in the middle of all
contributing timesteps of infile. This can be change with the CDO option
--timestat_date \<first|middle|last\>.
}
\details{
\preformatted{ daymin    Daily minimum
           For every adjacent sequence t_1, ...,t_n of timesteps of the same day it is:
}
}
