% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivepar.R
\name{Derivepar}
\alias{Derivepar}
\alias{cdo_sealevelpressure}
\title{Sea level pressure}
\usage{
cdo_sealevelpressure(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This operator computes the sea level pressure (air_pressure_at_sea_level).
Required input fields are surface_air_pressure, surface_geopotential and
air_temperature on hybrid sigma pressure levels.
}
\details{
\preformatted{ uvDestag      Destaggering of u/v wind components
               This is a special operator for destaggering of wind components.
               If the file contains a grid with temperature (name='t' or code=11)
               then grid_temp will be used for destaggered wind.
 rotuvNorth    Rotate u/v wind to North pole.
               This is an operator for transformation of wind-vectors from grid-relative to north-pole
               relative for the whole file. (FAST implementation with JACOBIANS)
 projuvLatLon  Cylindrical Equidistant projection
               Thus is an operator for transformation of wind-vectors from the globe-spherical coordinate system
               into a flat Cylindrical Equidistant (lat-lon) projection. (FAST JACOBIAN implementation)
}
}
