% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{Detrend}
\alias{Detrend}
\alias{cdo_detrend}
\title{Detrend time series}
\usage{
cdo_detrend(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Every time series in infile is linearly detrended. For every field element x
only those timesteps t belong to the sample S(x), which have i(t,x) NE miss. It
is assumed that all timesteps are equidistant, if this is not the case set the
parameter equal=false.
}
\details{
\preformatted{ ydrunmin   Multi-year daily running minimum
            o(001,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x); day[(i(t+(nts-1)/2)] = 001\}
}
}
