% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{Diff}
\alias{Diff}
\alias{cdo_diff}
\alias{cdo_diffn}
\title{Compare two datasets field by field}
\usage{
cdo_diff(infile1, infile2)

cdo_diff(infile1, infile2)

cdo_diffn(infile1, infile2)

cdo_diffn(infile1, infile2)
}
\arguments{
\item{infile1, infile2}{input files.}
}
\description{
Compares the contents of two datasets field by field. The input datasets need
to have the same structure and its fields need to have the same header
information and dimensions. Try the option names if the number of variables
differ. Exit status is 0 if inputs are the same and 1 if they differ.
}
\details{
\preformatted{ diff   Compare two datasets listed by parameter id
        Provides statistics on differences between two datasets.
        For each pair of fields the operator prints one line with the following information:
        - Date and Time
        - Level, Gridsize and number of Missing values
        - Number of different values
        - Occurrence of coefficient pairs with different signs (S)
        - Occurrence of zero values (Z)
        - Maxima of absolute difference of coefficient pairs
        - Maxima of relative difference of non-zero coefficient pairs with equal signs
        - Parameter identifier
 diffn  Compare two datasets listed by parameter name
        The same as operator diff. Using the name instead of the
        identifier to label the parameter.
}
}
