% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-build_functions.R
\name{.build_command}
\alias{.build_command}
\title{Build the command to run}
\usage{
.build_command(operator, inputs, outputs = NULL)
}
\arguments{
\item{operator}{name of the operator}

\item{inputs}{vector of inputs}

\item{outputs}{vector of outputs}
}
\description{
Takes the name of the operator and its arguments and build the command.
It doesn't run it unless the output is \code{""} (that is, the operator has
zero outputs)
}
