% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-build_functions.R
\name{.enchain}
\alias{.enchain}
\title{Prepare arguments for chaining}
\usage{
.enchain(input)
}
\arguments{
\item{input}{possible an operator.}
}
\description{
If \code{input} is the result of a cdo operation, then append "-".
}
\keyword{internal}
