% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-build_functions.R
\name{.new_step}
\alias{.new_step}
\title{Creates the structure of a cdo operation}
\usage{
.new_step(
  operator,
  inputs,
  outputs = NULL,
  n_inputs,
  n_outputs,
  params = list()
)
}
\arguments{
\item{operator}{name of the operator}

\item{inputs}{list of inputs}

\item{outputs}{outputs}

\item{n_inputs, n_outputs}{expected number of inputs and outputs.}

\item{params}{list of parameters (ignored)}

\item{obase}{base for outputs (for operators with -1 n_outputs)}
}
\description{
Creates the structure of a cdo operation
}
