% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate.R
\name{Duplicate}
\alias{Duplicate}
\alias{cdo_duplicate}
\title{Duplicates a dataset}
\usage{
cdo_duplicate(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This operator duplicates the contents of infile and writes the result to
outfile. The optional parameter sets the number of duplicates, the default is
2.
}
\details{
\preformatted{ copy  Copy datasets
       Copies all input datasets to outfile. 
 cat   Concatenate datasets
       Concatenates all input datasets and appends the result to the end 
       of outfile. If outfile does not exist it will be created.
}
}
