% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecacfd.R
\name{EcaCfd}
\alias{EcaCfd}
\alias{cdo_eca_cfd}
\title{Consecutive frost days index per time period}
\usage{
cdo_eca_cfd(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily minimum temperature TN, then the
largest number of consecutive days where TN \< 0 °C is counted. Note that TN
have to be given in units of Kelvin. A further output variable is the number of
frost periods of more than N days. The date information of a timestep in
outfile is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
