% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecacsu.R
\name{EcaCsu}
\alias{EcaCsu}
\alias{cdo_eca_csu}
\title{Consecutive summer days index per time period}
\usage{
cdo_eca_csu(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily maximum temperature TX, then the
largest number of consecutive days where TX \> T is counted. The number T is an
optional parameter with default T = 25°C. Note that TN have to be given in
units of Kelvin, whereas T have to be given in degrees Celsius. A further
output variable is the number of summer periods of more than N days. The date
information of a timestep in outfile is the date of the last contributing
timestep in infile.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
