% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecacwd.R
\name{EcaCwd}
\alias{EcaCwd}
\alias{cdo_eca_cwd}
\title{Consecutive wet days index per time period}
\usage{
cdo_eca_cwd(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Let infile be a time series of the daily precipitation amount RR, then the
largest number of consecutive days where RR is at least R is counted. R is an
optional parameter with default R = 1 mm. A further output variable is the
number of wet periods of more than N days. The date information of a timestep
in outfile is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
