% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecaetr.R
\name{EcaEtr}
\alias{EcaEtr}
\alias{cdo_eca_etr}
\title{Intra-period extreme temperature range}
\usage{
cdo_eca_etr(input_file_1, input_file_2, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2}{input files.}

\item{output_file}{output file.}
}
\description{
Let infile1 and infile2 be time series of thr maximum and minimum temperature
TX and TN, respectively. Then the extreme temperature range is the difference
of the maximum of TX and the minimum of TN. Note that TX and TN have to be
given in the same units. The date information of a timestep in outfile is the
date of the last contributing timesteps in infile1 and infile2.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
