% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecafd.R
\name{EcaFd}
\alias{EcaFd}
\alias{cdo_eca_fd}
\title{Frost days index per time period}
\usage{
cdo_eca_fd(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily minimum temperature TN, then the
number of days where TN \< 0 °C is counted. Note that TN have to be given in
units of Kelvin. The date information of a timestep in outfile is the date of
the last contributing timestep in infile.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
