% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecagsl.R
\name{EcaGsl}
\alias{EcaGsl}
\alias{cdo_eca_gsl}
\title{Thermal Growing season length index}
\usage{
cdo_eca_gsl(infile1, infile2, outfile = NULL)

cdo_eca_gsl(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
Let infile1 be a time series of the daily mean temperature TG, and infile2 be a
land-water mask. Within a period of 12 months, the thermal growing season
length is officially defined as the number of days between: - first occurrence
of at least nday consecutive days with TG \> T - first occurrence of at least
nday consecutive days with TG \< T within the last 6 months On northern
hemisphere, this period corresponds with the regular year, whereas on southern
hemisphere, it starts at July 1st. Please note, that this definition may lead
to weird results concerning values TG = T: In the first half of the period,
these days do not contribute to the gsl, but they do within the second half.
Moreover this definition could lead to discontinuous values in equatorial
regions.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
