% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecapd.R
\name{EcaPd}
\alias{EcaPd}
\alias{cdo_eca_pd}
\alias{cdo_eca_r10mm}
\alias{cdo_eca_r20mm}
\title{Precipitation days index per time period}
\usage{
cdo_eca_pd(infile, outfile = NULL)

cdo_eca_r10mm(infile, outfile = NULL)

cdo_eca_r20mm(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Let infile be a time series of the daily precipitation amount RR in \link{mm} (or
alternatively in \link{kg m-2}), then the number of days where RR is at least x mm
is counted. eca_r10mm and eca_r20mm are specific ECA operators with a daily
precipitation amount of 10 and 20 mm respectively. The date information of a
timestep in outfile is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
