% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecar99p.R
\name{EcaR99p}
\alias{EcaR99p}
\alias{cdo_eca_r99p}
\title{Extremely wet days wrt 99th percentile of reference period}
\usage{
cdo_eca_r99p(input_file_1, input_file_2, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2}{input files.}

\item{output_file}{output file.}
}
\description{
Let infile1 be a time series RR of the daily precipitation amount at wet days
(precipitation \>= 1 mm) and infile2 be the 99th percentile RRn99 of the daily
precipitation amount at wet days for any period used as reference. Then the
percentage of wet days with RR \> RRn99 is calculated. RRn99 is calculated as
the 99th percentile of all wet days of a given climate reference period.
Usually infile2 is generated by the operator ydaypctl,99. The date information
of a timestep in outfile is the date of the last contributing timestep in
infile1.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
