% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecarx1day.R
\name{EcaRx1day}
\alias{EcaRx1day}
\alias{cdo_eca_rx1day}
\title{Highest one day precipitation amount per time period}
\usage{
cdo_eca_rx1day(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily precipitation amount RR, then the
maximum of RR is written to outfile. If the optional parameter mode is set to
'm' the maximum daily precipitation amounts are determined for each month. The
date information of a timestep in outfile is the date of the last contributing
timestep in infile.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
