% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecarx5day.R
\name{EcaRx5day}
\alias{EcaRx5day}
\alias{cdo_eca_rx5day}
\title{Highest five-day precipitation amount per time period}
\usage{
cdo_eca_rx5day(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Let infile be a time series of 5-day precipitation totals RR, then the maximum
of RR is written to outfile. A further output variable is the number of 5 day
period with precipitation totals greater than x mm, where x is an optional
parameter with default x = 50 mm. The date information of a timestep in outfile
is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
