% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecasdii.R
\name{EcaSdii}
\alias{EcaSdii}
\alias{cdo_eca_sdii}
\title{Simple daily intensity index per time period}
\usage{
cdo_eca_sdii(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Let infile be a time series of the daily precipitation amount RR, then the mean
precipitation amount at wet days (RR \> R) is written to outfile. R is an
optional parameter with default R = 1 mm. The date information of a timestep in
outfile is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
