% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecasu.R
\name{EcaSu}
\alias{EcaSu}
\alias{cdo_eca_su}
\title{Summer days index per time period}
\usage{
cdo_eca_su(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily maximum temperature TX, then the
number of days where TX \> T is counted. The number T is an optional parameter
with default T = 25°C. Note that TX have to be given in units of Kelvin,
whereas T have to be given in degrees Celsius. The date information of a
timestep in outfile is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
