% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecatg90p.R
\name{EcaTg90p}
\alias{EcaTg90p}
\alias{cdo_eca_tg90p}
\title{Warm days percent wrt 90th percentile of reference period}
\usage{
cdo_eca_tg90p(infile1, infile2, outfile = NULL)

cdo_eca_tg90p(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
Let infile1 be a time series of the daily mean temperature TG, and infile2 be
the 90th percentile TGn90 of daily mean temperatures for any period used as
reference. Then the percentage of time where TG \> TGn90 is calculated. TGn90
is calculated as the 90th percentile of daily mean temperatures of a five day
window centred on each calendar day of a given climate reference period. Note
that both TG and TGn90 have to be given in the same units. The date information
of a timestep in outfile is the date of the last contributing timestep in
infile1.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
