% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecatn10p.R
\name{EcaTn10p}
\alias{EcaTn10p}
\alias{cdo_eca_tn10p}
\title{Cold nights percent wrt 10th percentile of reference period}
\usage{
cdo_eca_tn10p(input_file_1, input_file_2, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2}{input files.}

\item{output_file}{output file.}
}
\description{
Let infile1 be a time serie of the daily minimum temperature TN, and infile2 be
the 10th percentile TNn10 of daily minimum temperatures for any period used as
reference. Then the percentage of time where TN \< TNn10 is calculated. TNn10
is calculated as the 10th percentile of daily minimum temperatures of a five
day window centred on each calendar day of a given climate reference period.
Note that both TN and TNn10 have to be given in the same units. The date
information of a timestep in outfile is the date of the last contributing
timestep in infile1.
}
\details{
\preformatted{ eca_pd     Precipitation days index per time period
            Generic ECA operator with daily precipitation sum exceeding x mm.
 eca_r10mm  Heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 10 mm.
 eca_r20mm  Very heavy precipitation days index per time period
            Specific ECA operator with daily precipitation sum exceeding 20 mm.
}
}
