% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensstat.R
\name{Ensstat}
\alias{Ensstat}
\alias{cdo_ensavg}
\alias{cdo_ensmax}
\alias{cdo_ensmean}
\alias{cdo_ensmin}
\alias{cdo_enspctl}
\alias{cdo_ensrange}
\alias{cdo_ensstd}
\alias{cdo_ensstd1}
\alias{cdo_enssum}
\alias{cdo_ensvar}
\alias{cdo_ensvar1}
\title{Statistical values over an ensemble}
\usage{
cdo_ensavg(input_files, output_file = NULL)

cdo_ensmax(input_files, output_file = NULL)

cdo_ensmean(input_files, output_file = NULL)

cdo_ensmin(input_files, output_file = NULL)

cdo_enspctl(input_files, output_file = NULL)

cdo_ensrange(input_files, output_file = NULL)

cdo_ensstd(input_files, output_file = NULL)

cdo_ensstd1(input_files, output_file = NULL)

cdo_enssum(input_files, output_file = NULL)

cdo_ensvar(input_files, output_file = NULL)

cdo_ensvar1(input_files, output_file = NULL)
}
\arguments{
\item{input_files}{input files.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values over an ensemble of input files.
Depending on the chosen operator the minimum, maximum, range, sum, average,
variance, standard deviation or a certain percentile over all input files is
written to outfile. All input files need to have the same structure with the
same variables. The date information of a timestep in outfile is the date of
the first input file.
}
\details{
\preformatted{ ensmin    Ensemble minimum
           o(t,x) = min\{i1(t,x), i2(t,x), ..., in(t,x)\}
 ensmax    Ensemble maximum
           o(t,x) = max\{i1(t,x), i2(t,x), ..., in(t,x)\}
 ensrange  Ensemble range
           o(t,x) = range\{i1(t,x), i2(t,x), ..., in(t,x)\}
 enssum    Ensemble sum
           o(t,x) = sum\{i1(t,x), i2(t,x), ..., in(t,x)\}
 ensmean   Ensemble mean
           o(t,x) = mean\{i1(t,x), i2(t,x), ..., in(t,x)\}
 ensavg    Ensemble average
           o(t,x) = avg\{i1(t,x), i2(t,x), ..., in(t,x)\}
 ensstd    Ensemble standard deviation
           Normalize by n.
}
}
