% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensstat2.R
\name{Ensstat2}
\alias{Ensstat2}
\alias{cdo_ensrkhistspace}
\alias{cdo_ensrkhisttime}
\alias{cdo_ensroc}
\title{Statistical values over an ensemble}
\usage{
cdo_ensrkhistspace(input_files, output_file = NULL)

cdo_ensrkhisttime(input_files, output_file = NULL)

cdo_ensroc(input_files, output_file = NULL)
}
\arguments{
\item{input_files}{input files.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values over the ensemble of ensfiles using
obsfile as a reference. Depending on the operator a ranked Histogram or a
roc-curve over all Ensembles ensfiles with reference to obsfile is written to
outfile. The date and grid information of a timestep in outfile is the date of
the first input file. Thus all input files are required to have the same
structure in terms of the gridsize, variable definitions and number of
timesteps.
}
\details{
\preformatted{ ensrkhistspace  Ranked Histogram averaged over time
 ensrkhisttime   Ranked Histogram averaged over space
 ensroc          Ensemble Receiver Operating characteristics
}
}
