% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensval.R
\name{Ensval}
\alias{Ensval}
\alias{cdo_ensbrs}
\alias{cdo_enscrps}
\title{Ensemble validation tools}
\usage{
cdo_ensbrs(input_files, output_file = NULL)

cdo_enscrps(input_files, output_file = NULL)
}
\arguments{
\item{input_files}{input files.}

\item{output_file}{output file.}
}
\description{
This module computes ensemble validation scores and their decomposition such as
the Brier and cumulative ranked probability score (CRPS). The first file is
used as a reference it can be a climatology, observation or reanalysis against
which the skill of the ensembles given in infiles is measured. Depending on the
operator a number of output files is generated each containing the skill score
and its decomposition corresponding to the operator. The output is averaged
over horizontal fields using appropriate weights for each level and timestep in
rfile.
}
\details{
\preformatted{ enscrps  Ensemble CRPS and decomposition
 ensbrs   Ensemble Brier score
          Ensemble Brier Score and Decomposition
}
}
