% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofcoeff.R
\name{Eofcoeff}
\alias{Eofcoeff}
\alias{cdo_eofcoeff}
\title{Principal coefficients of EOFs}
\usage{
cdo_eofcoeff(infile1, infile2, obase = NULL)

cdo_eofcoeff(infile1, infile2, obase = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{obase}{preffix added to the output files.}
}
\description{
This module calculates the time series of the principal coefficients for given
EOF (empirical orthogonal functions) and data. Time steps in infile1 are
assumed to be the EOFs, time steps in infile2 are assumed to be the time
series. Note, that this operator calculates a non weighted dot product of the
fields in infile1 and infile2. For consistency set the environment variable
CDO_WEIGHT_MODE=off when using eof or eof3d.
}
\details{
\preformatted{ eof         Calculate EOFs in spatial or time space
 eoftime     Calculate EOFs in time space
 eofspatial  Calculate EOFs in spatial space
 eof3d       Calculate 3-Dimensional EOFs in time space
}
}
