% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofs.R
\name{EOFs}
\alias{EOFs}
\alias{cdo_eof}
\alias{cdo_eof3d}
\alias{cdo_eofspatial}
\alias{cdo_eoftime}
\title{Empirical Orthogonal Functions}
\usage{
cdo_eof(input_file, output_file_1 = NULL, output_file_2 = NULL)

cdo_eof3d(input_file, output_file_1 = NULL, output_file_2 = NULL)

cdo_eofspatial(input_file, output_file_1 = NULL, output_file_2 = NULL)

cdo_eoftime(input_file, output_file_1 = NULL, output_file_2 = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file_1, output_file_2}{output files.}
}
\description{
This module calculates empirical orthogonal functions of the data in infile as
the eigen values of the scatter matrix (covariance matrix) S of the data sample
z(t). A more detailed description can be found above.
}
\details{
\preformatted{ eof         Calculate EOFs in spatial or time space
 eoftime     Calculate EOFs in time space
 eofspatial  Calculate EOFs in spatial space
 eof3d       Calculate 3-Dimensional EOFs in time space
}
}
