% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{Expr}
\alias{Expr}
\alias{cdo_aexpr}
\alias{cdo_aexprf}
\alias{cdo_expr}
\alias{cdo_exprf}
\title{Evaluate expressions}
\usage{
cdo_aexpr(input_file, output_file = NULL)

cdo_aexprf(input_file, output_file = NULL)

cdo_expr(input_file, output_file = NULL)

cdo_exprf(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module arithmetically processes every timestep of the input dataset. Each
individual assignment statement have to end with a semi-colon. Unlike regular
variables, temporary variables are never written to the output stream. To
define a temporary variable simply prefix the variable name with an underscore
(e.g. _varname) when the variable is declared.
}
\details{
\preformatted{ expr    Evaluate expressions
         The processing instructions are read from the parameter.
 exprf   Evaluate expressions script
         Contrary to expr the processing instructions are read from a file.
 aexpr   Evaluate expressions and append results
         Same as expr, but keep input variables and append results
 aexprf  Evaluate expression script and append results
         Same as exprf, but keep input variables and append results
}
}
