% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedes.R
\name{Filedes}
\alias{Filedes}
\alias{cdo_codetab}
\alias{cdo_griddes}
\alias{cdo_partab}
\alias{cdo_vct}
\alias{cdo_zaxisdes}
\title{Dataset description}
\usage{
cdo_codetab(input_file)

cdo_griddes(input_file)

cdo_partab(input_file)

cdo_vct(input_file)

cdo_zaxisdes(input_file)
}
\arguments{
\item{input_file}{input file.}
}
\description{
This module provides operators to print meta information about a dataset. The
printed meta-data depends on the chosen operator.
}
\details{
\preformatted{ partab    Parameter table
           Prints all available meta information of the variables.
 codetab   Parameter code table
           Prints a code table with a description of all variables.
           For each variable the operator prints one line listing the
           code, name, description and units.
 griddes   Grid description
           Prints the description of all grids.
 zaxisdes  Z-axis description
           Prints the description of all z-axes.
 vct       Vertical coordinate table
           Prints the vertical coordinate table.
}
}
