% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldcovar.R
\name{Fldcovar}
\alias{Fldcovar}
\alias{cdo_fldcovar}
\title{Covariance in grid space}
\usage{
cdo_fldcovar(infile1, infile2, outfile = NULL)

cdo_fldcovar(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This operator calculates the covariance of two fields over all gridpoints for
each timestep. With
}
\details{
\preformatted{ ydrunmin   Multi-year daily running minimum
            o(001,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x); day[(i(t+(nts-1)/2)] = 001\}
}
}
