% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldstat.R
\name{Fldstat}
\alias{Fldstat}
\alias{cdo_fldavg}
\alias{cdo_fldmax}
\alias{cdo_fldmean}
\alias{cdo_fldmin}
\alias{cdo_fldpctl}
\alias{cdo_fldrange}
\alias{cdo_fldstd}
\alias{cdo_fldstd1}
\alias{cdo_fldsum}
\alias{cdo_fldvar}
\alias{cdo_fldvar1}
\title{Statistical values over a field}
\usage{
cdo_fldavg(input_file, output_file = NULL)

cdo_fldmax(input_file, output_file = NULL)

cdo_fldmean(input_file, output_file = NULL)

cdo_fldmin(input_file, output_file = NULL)

cdo_fldpctl(input_file, output_file = NULL)

cdo_fldrange(input_file, output_file = NULL)

cdo_fldstd(input_file, output_file = NULL)

cdo_fldstd1(input_file, output_file = NULL)

cdo_fldsum(input_file, output_file = NULL)

cdo_fldvar(input_file, output_file = NULL)

cdo_fldvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values of all input fields. A field is a
horizontal layer of a data variable. According to the chosen operator the field
minimum, maximum, range, sum, average, variance, standard deviation or a
certain percentile is written to outfile.
}
\details{
\preformatted{ fldmin    Field minimum
           For every gridpoint x_1, ..., x_n of the same field it is:
}
}
