% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{Fourier}
\alias{Fourier}
\alias{cdo_fourier}
\title{Fourier transformation}
\usage{
cdo_fourier(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
The fourier operator performs the fourier transformation or the inverse fourier
transformation of all input fields. If the number of timesteps is a power of 2
then the algorithm of the Fast Fourier Transformation (FFT) is used.
}
\details{
\preformatted{ dv2uv  Divergence and vorticity to U and V wind
        Calculate U and V wind on a Gaussian grid from spherical harmonic 
        coefficients of relative divergence and vorticity. The divergence and vorticity 
        need to have the names sd and svo or code numbers 155 and 138.
 uv2dv  U and V wind to divergence and vorticity
        Calculate spherical harmonic coefficients of relative divergence and vorticity
        from U and V wind. The U and V wind need to be on a Gaussian grid and need to have the 
        names u and v or the code numbers 131 and 132.
}
}
