% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradsdes.R
\name{Gradsdes}
\alias{Gradsdes}
\alias{cdo_gradsdes}
\title{GrADS data descriptor file}
\usage{
cdo_gradsdes(input_file)
}
\arguments{
\item{input_file}{input file.}
}
\description{
Creates a GrADS data descriptor file. Supported file formats are GRIB1, NetCDF,
SERVICE, EXTRA and IEG. For GRIB1 files the GrADS map file is also generated.
For SERVICE and EXTRA files the grid have to be specified with the CDO option
'-g \<grid\>'. This module takes infile in order to create filenames for the
descriptor (infile.ctl) and the map (infile.gmp) file.
}
\details{
\preformatted{ gmtxyz    GMT xyz format
           The operator exports the first field to the GMT xyz ASCII format.
           The output can be used to create contour plots with the GMT module pscontour.
 gmtcells  GMT multiple segment format
           The operator exports the first field to the GMT multiple segment ASCII format.
           The output can be used to create shaded gridfill plots with the GMT module psxy.
}
}
