% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcell.R
\name{Gridcell}
\alias{Gridcell}
\alias{cdo_gridarea}
\alias{cdo_gridweights}
\title{Grid cell quantities}
\usage{
cdo_gridarea(infile, outfile = NULL)

cdo_gridweights(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module reads the grid cell area of the first grid from the input stream.
If the grid cell area is missing it will be computed from the grid description.
Depending on the chosen operator the grid cell area or weights are written to
the output stream.
}
\details{
\preformatted{ gridarea     Grid cell area
              Writes the grid cell area to the output stream. If the grid cell area have to
              be computed it is scaled with the earth radius to square meters.
 gridweights  Grid cell weights
              Writes the grid cell area weights to the output stream.
}
}
