% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{Histogram}
\alias{Histogram}
\alias{cdo_histcount}
\alias{cdo_histfreq}
\alias{cdo_histmean}
\alias{cdo_histsum}
\title{Histogram}
\usage{
cdo_histcount(infile, outfile = NULL)

cdo_histfreq(infile, outfile = NULL)

cdo_histmean(infile, outfile = NULL)

cdo_histsum(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module creates bins for a histogram of the input data. The bins have to be
adjacent and have non-overlapping intervals. The user has to define the bounds
of the bins. The first value is the lower bound and the second value the upper
bound of the first bin. The bounds of the second bin are defined by the second
and third value, aso. Only 2-dimensional input fields are allowed. The output
file contains one vertical level for each of the bins requested.
}
\details{
\preformatted{ histcount  Histogram count
            Number of elements in the bin range.
 histsum    Histogram sum
            Sum of elements in the bin range.
 histmean   Histogram mean
            Mean of elements in the bin range.
 histfreq   Histogram frequency
            Relative frequency of elements in the bin range.
}
}
