% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourstat.R
\name{Hourstat}
\alias{Hourstat}
\alias{cdo_houravg}
\alias{cdo_hourmax}
\alias{cdo_hourmean}
\alias{cdo_hourmin}
\alias{cdo_hourrange}
\alias{cdo_hourstd}
\alias{cdo_hourstd1}
\alias{cdo_hoursum}
\alias{cdo_hourvar}
\alias{cdo_hourvar1}
\title{Hourly statistical values}
\usage{
cdo_houravg(infile, outfile = NULL)

cdo_hourmax(infile, outfile = NULL)

cdo_hourmean(infile, outfile = NULL)

cdo_hourmin(infile, outfile = NULL)

cdo_hourrange(infile, outfile = NULL)

cdo_hourstd(infile, outfile = NULL)

cdo_hourstd1(infile, outfile = NULL)

cdo_hoursum(infile, outfile = NULL)

cdo_hourvar(infile, outfile = NULL)

cdo_hourvar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes statistical values over timesteps of the same hour.
Depending on the chosen operator the minimum, maximum, range, sum, average,
variance or standard deviation of timesteps of the same hour is written to
outfile. The time of outfile is determined by the time in the middle of all
contributing timesteps of infile. This can be change with the CDO option
--timestat_date \<first|middle|last\>.
}
\details{
\preformatted{ hourmin    Hourly minimum
            For every adjacent sequence t_1, ...,t_n of timesteps of the same hour it is:
}
}
