% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importamsr.R
\name{Importamsr}
\alias{Importamsr}
\alias{cdo_import_amsr}
\title{Import AMSR binary files}
\usage{
cdo_import_amsr(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This operator imports gridded binary AMSR (Advanced Microwave Scanning
Radiometer) data. The binary data files are available from the AMSR ftp site
(ftp://ftp.ssmi.com/amsre). Each file consists of twelve (daily) or five
(averaged) 0.25 x 0.25 degree grid (1440,720) byte maps. For daily files, six
daytime maps in the following order, Time (UTC), Sea Surface Temperature (SST),
10 meter Surface Wind Speed (WSPD), Atmospheric Water Vapor (VAPOR), Cloud
Liquid Water (CLOUD), and Rain Rate (RAIN), are followed by six nighttime maps
in the same order. Time-Averaged files contain just the geophysical layers in
the same order \link{SST, WSPD, VAPOR, CLOUD, RAIN}. More information to the data is
available on the AMSR homepage http://www.remss.com/amsr.
}
\details{
\preformatted{ dv2uv  Divergence and vorticity to U and V wind
        Calculate U and V wind on a Gaussian grid from spherical harmonic 
        coefficients of relative divergence and vorticity. The divergence and vorticity 
        need to have the names sd and svo or code numbers 155 and 138.
 uv2dv  U and V wind to divergence and vorticity
        Calculate spherical harmonic coefficients of relative divergence and vorticity
        from U and V wind. The U and V wind need to be on a Gaussian grid and need to have the 
        names u and v or the code numbers 131 and 132.
}
}
