% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intlevel.R
\name{Intlevel}
\alias{Intlevel}
\alias{cdo_intlevel}
\title{Linear level interpolation}
\usage{
cdo_intlevel(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This operator performs a linear vertical interpolation of non hybrid 3D
variables.
}
\details{
\preformatted{ ap2pl  Air pressure to pressure level interpolation
        Interpolates 3D variables on hybrid sigma height coordinates to pressure level.
 ap2hl  Air pressure to height level interpolation
        Interpolates 3D variables on hybrid sigma height coordinates to height level.
        The procedure is the same as for the operator ap2pl except for
        the pressure levels being calculated from the heights by:
        plevel = 101325*exp(hlevel/-7000)
}
}
