% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intlevel3d.R
\name{Intlevel3d}
\alias{Intlevel3d}
\alias{cdo_intlevel3d}
\alias{cdo_intlevelx3d}
\title{Linear level interpolation from/to 3d vertical coordinates}
\usage{
cdo_intlevel3d(input_file_1, input_file_2, output_file = NULL)

cdo_intlevelx3d(input_file_1, input_file_2, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2}{input files.}

\item{output_file}{output file.}
}
\description{
This operator performs a linear vertical interpolation of 3D variables fields
with given 3D vertical coordinates.
}
\details{
\preformatted{ intlevel3d   Linear level interpolation onto a 3d vertical coordinate
 intlevelx3d  like intlevel3d but with extrapolation
}
}
