% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertlev.R
\name{Invertlev}
\alias{Invertlev}
\alias{cdo_invertlev}
\title{Invert levels}
\usage{
cdo_invertlev(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This operator inverts the levels of all 3D variables.
}
\details{
\preformatted{ setzaxis        Set z-axis
                 This operator sets the z-axis description of all variables with the same number of level as the new z-axis.
 genlevelbounds  Generate level bounds
                 Generates the layer bounds of the z-axis.
}
}
