% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maggraph.R
\name{Maggraph}
\alias{Maggraph}
\alias{cdo_graph}
\title{Line graph plot}
\usage{
cdo_graph(infiles, outfile = NULL)
}
\arguments{
\item{infiles}{input files.}

\item{outfile}{output file.}
}
\description{
This operator generates line graph plots. The data for the plot is read from
infiles. The result is written to outfile. The default output file format is
postscript, this can be changed with the device parameter.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
