% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magplot.R
\name{Magplot}
\alias{Magplot}
\alias{cdo_contour}
\alias{cdo_grfill}
\alias{cdo_shaded}
\title{Lat/Lon plot}
\usage{
cdo_contour(infile, outfile = NULL)

cdo_grfill(infile, outfile = NULL)

cdo_shaded(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
The operators in this module generates 2D Lon/Lat plots. The data for the plot
is read from infile. Only data on rectilinear Lon/Lat grids are supported. The
output file will be named \<obase\>_\<param\>.\<device\> where param is the
parameter name and device is the device name. The default output file format is
postscript, this can be changed with the device parameter. The type of the plot
depends on the choosen operator.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
