% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magvector.R
\name{Magvector}
\alias{Magvector}
\alias{cdo_vector}
\title{Lat/Lon vector plot}
\usage{
cdo_vector(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This operator generates 2D Lon/Lat vector plots. The data for the plot is read
from infile. The input is expected to contain two velocity components. Only
data on rectilinear Lon/Lat grids are supported. The output file will be named
\<obase\>.\<device\> where device is the device name. The default output file
format is postscript, this can be changed with the device parameter.
}
\details{
\preformatted{ contour  Contour plot
          The operator contour generates the discrete contour lines of the input field values.
          The following additional parameters are valid for contour operator,
          module in addition to the common plot parameters:
}
}
