% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskbox.R
\name{Maskbox}
\alias{Maskbox}
\alias{cdo_maskindexbox}
\alias{cdo_masklonlatbox}
\title{Mask a box}
\usage{
cdo_maskindexbox(input_file, output_file = NULL)

cdo_masklonlatbox(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Masked a box of the rectangularly understood field. The elements inside the box
are untouched, the elements outside are set to missing value. All input fields
need to have the same horizontal grid. Use sellonlatbox or selindexbox if only
the data inside the box are needed.
}
\details{
\preformatted{ masklonlatbox  Mask a longitude/latitude box
                Masked a regular longitude/latitude box. The user has to give the longitudes and latitudes of the 
                edges of the box. Considered are only those grid cells with the grid center inside the lon/lat box.
 maskindexbox   Mask an index box
                Masked an index box. The user has to give the indexes of the edges of the box. 
                The index of the left edge can be greater then the one of the right edge.
}
}
