% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskregion.R
\name{Maskregion}
\alias{Maskregion}
\alias{cdo_maskregion}
\title{Mask regions}
\usage{
cdo_maskregion(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Masks different regions of fields with a regular lon/lat grid. The elements
inside a region are untouched, the elements outside are set to missing value.
Considered are only those grid cells with the grid center inside the regions.
All input fields must have the same horizontal grid. The user has to give ASCII
formatted files with different regions. A region is defined by a polygon. Each
line of a polygon description file contains the longitude and latitude of one
point. Each polygon description file can contain one or more polygons separated
by a line with the character \\&.
}
\details{
\preformatted{ shiftx  Shift x
         Shifts all fields in x direction.
 shifty  Shift y
         Shifts all fields in y direction.
}
}
