% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merstat.R
\name{Merstat}
\alias{Merstat}
\alias{cdo_meravg}
\alias{cdo_mermax}
\alias{cdo_mermean}
\alias{cdo_mermin}
\alias{cdo_merpctl}
\alias{cdo_merrange}
\alias{cdo_merstd}
\alias{cdo_merstd1}
\alias{cdo_mersum}
\alias{cdo_mervar}
\alias{cdo_mervar1}
\title{Meridional statistical values}
\usage{
cdo_meravg(infile, outfile = NULL)

cdo_mermax(infile, outfile = NULL)

cdo_mermean(infile, outfile = NULL)

cdo_mermin(infile, outfile = NULL)

cdo_merpctl(infile, outfile = NULL)

cdo_merrange(infile, outfile = NULL)

cdo_merstd(infile, outfile = NULL)

cdo_merstd1(infile, outfile = NULL)

cdo_mersum(infile, outfile = NULL)

cdo_mervar(infile, outfile = NULL)

cdo_mervar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes meridional statistical values of the input fields.
According to the chosen operator the meridional minimum, maximum, range, sum,
average, variance, standard deviation or a certain percentile is written to
outfile. This operator requires all variables on the same regular lon/lat grid.
}
\details{
\preformatted{ mermin    Meridional minimum
           For every longitude the minimum over all latitudes is computed.
 mermax    Meridional maximum
           For every longitude the maximum over all latitudes is computed.
 merrange  Meridional range
           For every longitude the range over all latitudes is computed.
 mersum    Meridional sum
           For every longitude the sum over all latitudes is computed.
 mermean   Meridional mean
           For every longitude the area weighted mean over all latitudes is computed.
 meravg    Meridional average
           For every longitude the area weighted average over all latitudes is computed.
 merstd    Meridional standard deviation
           For every longitude the standard deviation over all latitudes is computed. Normalize by n.
 merstd1   Meridional standard deviation (n-1)
           For every longitude the standard deviation over all latitudes is computed. Normalize by (n-1).
 mervar    Meridional variance
           For every longitude the variance over all latitudes is computed. Normalize by n.
 mervar1   Meridional variance (n-1)
           For every longitude the variance over all latitudes is computed. Normalize by (n-1).
 merpctl   Meridional percentiles
           For every longitude the pth percentile over all latitudes is computed.
}
}
