% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarith.R
\name{Monarith}
\alias{Monarith}
\alias{cdo_monadd}
\alias{cdo_mondiv}
\alias{cdo_monmul}
\alias{cdo_monsub}
\title{Monthly arithmetic}
\usage{
cdo_monadd(infile1, infile2, outfile = NULL)

cdo_monadd(infile1, infile2, outfile = NULL)

cdo_mondiv(infile1, infile2, outfile = NULL)

cdo_mondiv(infile1, infile2, outfile = NULL)

cdo_monmul(infile1, infile2, outfile = NULL)

cdo_monmul(infile1, infile2, outfile = NULL)

cdo_monsub(infile1, infile2, outfile = NULL)

cdo_monsub(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This module performs simple arithmetic of a time series and one timestep with
the same month and year. For each field in infile1 the corresponding field of
the timestep in infile2 with the same month and year is used. The header
information in infile1 have to be the same as in infile2. Usually infile2 is
generated by an operator of the module MONSTAT.
}
\details{
\preformatted{ monadd  Add monthly time series
         Adds a time series and a monthly time series.
 monsub  Subtract monthly time series
         Subtracts a time series and a monthly time series.
 monmul  Multiply monthly time series
         Multiplies a time series and a monthly time series.
 mondiv  Divide monthly time series
         Divides a time series and a monthly time series.
}
}
