% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_wind.R
\name{NCL_wind}
\alias{NCL_wind}
\alias{cdo_uv2dv_cfd}
\alias{cdo_uv2vr_cfd}
\title{Wind transformation}
\usage{
cdo_uv2dv_cfd(input_file, output_file = NULL)

cdo_uv2vr_cfd(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module contains CDO operators with an interface to NCL functions. The
corresponding NCL functions have the same name. A more detailed description of
those NCL function can be found on the NCL homepage https://www.ncl.ucar.edu.
}
\details{
\preformatted{ uv2vr_cfd  U and V wind to relative vorticity
            Computes relative vorticity for a latitude-longitude grid using centered finite differences.
            The grid need not be global and missing values are allowed.
 uv2dv_cfd  U and V wind to divergence
            Computes divergence for a latitude-longitude grid using centered finite differences.
            The grid need not be global and missing values are allowed.
}
}
