% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-build_functions.R
\name{new_operator}
\alias{new_operator}
\title{Create an operator}
\usage{
new_operator(operator, n_inputs, n_outputs)
}
\arguments{
\item{operator}{name of the operator.}

\item{n_inputs}{number of inputs (-1 means arbitrary number of inputs)}

\item{n_outputs}{number of outputs}
}
\description{
Creates an operator out of its specifications
}
