% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputgmt.R
\name{Outputgmt}
\alias{Outputgmt}
\alias{cdo_gmtcells}
\alias{cdo_gmtxyz}
\title{GMT output}
\usage{
cdo_gmtcells(input_file)

cdo_gmtxyz(input_file)
}
\arguments{
\item{input_file}{input file.}
}
\description{
This module prints the first field of the input dataset to standard output. The
output can be used to generate 2D Lon/Lat plots with GMT. The format of the
output depends on the chosen operator.
}
\details{
\preformatted{ gmtxyz    GMT xyz format
           The operator exports the first field to the GMT xyz ASCII format.
           The output can be used to create contour plots with the GMT module pscontour.
 gmtcells  GMT multiple segment format
           The operator exports the first field to the GMT multiple segment ASCII format.
           The output can be used to create shaded gridfill plots with the GMT module psxy.
}
}
