% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapbic.R
\name{Remapbic}
\alias{Remapbic}
\alias{cdo_genbic}
\alias{cdo_remapbic}
\title{Bicubic interpolation}
\usage{
cdo_genbic(infile, outfile = NULL)

cdo_remapbic(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module contains operators for a bicubic remapping of fields between grids
in spherical coordinates. The interpolation is based on an adapted SCRIP
library version. For a detailed description of the interpolation method see
SCRIP. This interpolation method only works on quadrilateral curvilinear source
grids.
}
\details{
\preformatted{ remapbic  Bicubic interpolation
           Performs a bicubic interpolation on all input fields.
 genbic    Generate bicubic interpolation weights
           Generates bicubic interpolation weights for the first input field and writes the
           result to a file. The format of this file is NetCDF following the SCRIP convention.
           Use the operator remap to apply this remapping weights to a data file with the same source grid.
}
}
