% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapbil.R
\name{Remapbil}
\alias{Remapbil}
\alias{cdo_genbil}
\alias{cdo_remapbil}
\title{Bilinear interpolation}
\usage{
cdo_genbil(infile, outfile = NULL)

cdo_remapbil(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module contains operators for a bilinear remapping of fields between grids
in spherical coordinates. The interpolation is based on an adapted SCRIP
library version. For a detailed description of the interpolation method see
SCRIP. This interpolation method only works on quadrilateral curvilinear source
grids.
}
\details{
\preformatted{ remapbil  Bilinear interpolation
           Performs a bilinear interpolation on all input fields.
 genbil    Generate bilinear interpolation weights
           Generates bilinear interpolation weights for the first input field and writes the
           result to a file. The format of this file is NetCDF following the SCRIP convention.
           Use the operator remap to apply this remapping weights to a data file with the same source grid.
}
}
