% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapcon.R
\name{Remapcon}
\alias{Remapcon}
\alias{cdo_gencon}
\alias{cdo_remapcon}
\title{First order conservative remapping}
\usage{
cdo_gencon(infile, outfile = NULL)

cdo_remapcon(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module contains operators for a first order conservative remapping of
fields between grids in spherical coordinates. The operators in this module
uses code from the YAC software package to compute the conservative remapping
weights. For a detailed description of the interpolation method see YAC. The
interpolation method is completely general and can be used for any grid on a
sphere. The search algorithm for the conservative remapping requires that no
grid cell occurs more than once.
}
\details{
\preformatted{ remapcon  First order conservative remapping
           Performs a first order conservative remapping on all input fields.
 gencon    Generate 1st order conservative remap weights
           Generates first order conservative remapping weights for the first input field and
           writes the result to a file. The format of this file is NetCDF following the SCRIP convention.
           Use the operator remap to apply this remapping weights to a data file with the same source grid.
}
}
