% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapcon2.R
\name{Remapcon2}
\alias{Remapcon2}
\alias{cdo_gencon2}
\alias{cdo_remapcon2}
\title{Second order conservative remapping}
\usage{
cdo_gencon2(infile, outfile = NULL)

cdo_remapcon2(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module contains operators for a second order conservative remapping of
fields between grids in spherical coordinates. The interpolation is based on an
adapted SCRIP library version. For a detailed description of the interpolation
method see SCRIP. The interpolation method is completely general and can be
used for any grid on a sphere. The search algorithm for the conservative
remapping requires that no grid cell occurs more than once.
}
\details{
\preformatted{ remapcon2  Second order conservative remapping
            Performs a second order conservative remapping on all input fields.
 gencon2    Generate 2nd order conservative remap weights
            Generates second order conservative remapping weights for the first input field and
            writes the result to a file. The format of this file is NetCDF following the SCRIP convention.
            Use the operator remap to apply this remapping weights to a data file with the same source grid.
}
}
