% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacevalues.R
\name{Replacevalues}
\alias{Replacevalues}
\alias{cdo_setrtoc}
\alias{cdo_setrtoc2}
\alias{cdo_setvals}
\title{Replace variable values}
\usage{
cdo_setrtoc(infile, outfile = NULL)

cdo_setrtoc2(infile, outfile = NULL)

cdo_setvals(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module replaces old variable values with new values, depending on the
operator.
}
\details{
\preformatted{ setvals   Set list of old values to new values
           Supply a list of n pairs of old and new values.
 setrtoc   Set range to constant
                    / c      if i(t,x) GE rmin AND i(t,x) LE rmax
           o(t,x) = 
                    \\\\ i(t,x) if i(t,x) LT rmin AND i(t,x) GT rmax
 setrtoc2  Set range to constant others to constant2
                    / c      if i(t,x) GE rmin AND i(t,x) LE rmax
           o(t,x) = 
                    \\\\ c2     if i(t,x) LT rmin AND i(t,x) GT rmax
}
}
