% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpctl.R
\name{Runpctl}
\alias{Runpctl}
\alias{cdo_runpctl}
\title{Running percentile values}
\usage{
cdo_runpctl(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes running percentiles over a selected number of timesteps in
infile. The time of outfile is determined by the time in the middle of all
contributing timesteps of infile. This can be change with the CDO option
--timestat_date \<first|middle|last\>.
}
\details{
\preformatted{ runmin    Running minimum
           o(t+(nts-1)/2,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runmax    Running maximum
           o(t+(nts-1)/2,x) = max\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runrange  Running range
           o(t+(nts-1)/2,x) = range\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runsum    Running sum
           o(t+(nts-1)/2,x) = sum\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runmean   Running mean
           o(t+(nts-1)/2,x) = mean\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runavg    Running average
           o(t+(nts-1)/2,x) = avg\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runstd    Running standard deviation
           Normalize by n. 
}
}
