% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstat.R
\name{Runstat}
\alias{Runstat}
\alias{cdo_runavg}
\alias{cdo_runmax}
\alias{cdo_runmean}
\alias{cdo_runmin}
\alias{cdo_runrange}
\alias{cdo_runstd}
\alias{cdo_runstd1}
\alias{cdo_runsum}
\alias{cdo_runvar}
\alias{cdo_runvar1}
\title{Running statistical values}
\usage{
cdo_runavg(infile, outfile = NULL)

cdo_runmax(infile, outfile = NULL)

cdo_runmean(infile, outfile = NULL)

cdo_runmin(infile, outfile = NULL)

cdo_runrange(infile, outfile = NULL)

cdo_runstd(infile, outfile = NULL)

cdo_runstd1(infile, outfile = NULL)

cdo_runsum(infile, outfile = NULL)

cdo_runvar(infile, outfile = NULL)

cdo_runvar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes running statistical values over a selected number of
timesteps. Depending on the chosen operator the minimum, maximum, range, sum,
average, variance or standard deviation of a selected number of consecutive
timesteps read from infile is written to outfile. The time of outfile is
determined by the time in the middle of all contributing timesteps of infile.
This can be change with the CDO option --timestat_date \<first|middle|last\>.
}
\details{
\preformatted{ runmin    Running minimum
           o(t+(nts-1)/2,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runmax    Running maximum
           o(t+(nts-1)/2,x) = max\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runrange  Running range
           o(t+(nts-1)/2,x) = range\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runsum    Running sum
           o(t+(nts-1)/2,x) = sum\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runmean   Running mean
           o(t+(nts-1)/2,x) = mean\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runavg    Running average
           o(t+(nts-1)/2,x) = avg\{i(t,x), i(t+1,x), ..., i(t+nts-1,x)\}
 runstd    Running standard deviation
           Normalize by n. 
}
}
