% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplegrid.R
\name{Samplegrid}
\alias{Samplegrid}
\alias{cdo_samplegrid}
\title{Resample grid}
\usage{
cdo_samplegrid(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This is a special operator for resampling the horizontal grid. No interpolation
takes place. Resample factor=2 means every second grid point is removed. Only
rectilinear and curvilinear source grids are supported by this operator.
}
\details{
\preformatted{ selgridcell  Select grid cells
 delgridcell  Delete grid cells
}
}
