% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasstat.R
\name{Seasstat}
\alias{Seasstat}
\alias{cdo_seasavg}
\alias{cdo_seasmax}
\alias{cdo_seasmean}
\alias{cdo_seasmin}
\alias{cdo_seasrange}
\alias{cdo_seasstd}
\alias{cdo_seasstd1}
\alias{cdo_seassum}
\alias{cdo_seasvar}
\alias{cdo_seasvar1}
\title{Seasonal statistical values}
\usage{
cdo_seasavg(infile, outfile = NULL)

cdo_seasmax(infile, outfile = NULL)

cdo_seasmean(infile, outfile = NULL)

cdo_seasmin(infile, outfile = NULL)

cdo_seasrange(infile, outfile = NULL)

cdo_seasstd(infile, outfile = NULL)

cdo_seasstd1(infile, outfile = NULL)

cdo_seassum(infile, outfile = NULL)

cdo_seasvar(infile, outfile = NULL)

cdo_seasvar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes statistical values over timesteps of the same season.
Depending on the chosen operator the minimum, maximum, range, sum, average,
variance or standard deviation of timesteps of the same season is written to
outfile. The time of outfile is determined by the time in the middle of all
contributing timesteps of infile. This can be change with the CDO option
--timestat_date \<first|middle|last\>. Be careful about the first and the last
output timestep, they may be incorrect values if the seasons have incomplete
timesteps.
}
\details{
\preformatted{ seasmin    Seasonal minimum
            For every adjacent sequence t_1, ...,t_n of timesteps of the same season it is:
}
}
