% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selbox.R
\name{Selbox}
\alias{Selbox}
\alias{cdo_selindexbox}
\alias{cdo_sellonlatbox}
\title{Select a box of a field}
\usage{
cdo_selindexbox(input_file, output_file = NULL)

cdo_sellonlatbox(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Selects a box of the rectangularly understood field.
}
\details{
\preformatted{ sellonlatbox  Select a longitude/latitude box
               Selects a regular longitude/latitude box. The user has to give the longitudes and latitudes of the 
               edges of the box. Considered are only those grid cells with the grid center inside the lon/lat box.
               For rotated lon/lat grids the parameter needs to be rotated coordinates.
 selindexbox   Select an index box
               Selects an index box. The user has to give the indexes of the edges of the box. The index of the 
               left edge may be greater then that of the right edge.
}
}
