% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{Select}
\alias{Select}
\alias{cdo_delete}
\alias{cdo_select}
\title{Select fields}
\usage{
cdo_delete(infiles, outfile = NULL)

cdo_select(infiles, outfile = NULL)
}
\arguments{
\item{infiles}{input files.}

\item{outfile}{output file.}
}
\description{
This module selects some fields from infiles and writes them to outfile.
infiles is an arbitrary number of input files. All input files need to have the
same structure with the same variables on different timesteps. The fields
selected depends on the chosen parameters. Parameter is a comma-separated list
of \"key=value\" pairs. A range of integer values can be specified by
first/last\link{/inc}. Wildcards can be used for string values.
}
\details{
\preformatted{ select  Select fields
         Selects all fields with parameters in a user given list.
 delete  Delete fields
         Deletes all fields with parameters in a user given list.
}
}
