% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selgridcell.R
\name{Selgridcell}
\alias{Selgridcell}
\alias{cdo_delgridcell}
\alias{cdo_selgridcell}
\title{Select grid cells}
\usage{
cdo_delgridcell(infile, outfile = NULL)

cdo_selgridcell(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Selects grid cells of all fields from infile. The user has to give the indexes
of each grid cell. The resulting grid in outfile is unstructured.
}
\details{
\preformatted{ selgridcell  Select grid cells
 delgridcell  Delete grid cells
}
}
