% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selmulti.R
\name{Selmulti}
\alias{Selmulti}
\alias{cdo_changemulti}
\alias{cdo_delmulti}
\alias{cdo_selmulti}
\title{Select multiple fields via GRIB1 parameters}
\usage{
cdo_changemulti(input_file, output_file = NULL)

cdo_delmulti(input_file, output_file = NULL)

cdo_selmulti(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module selects multiple fields from infile and writes them to outfile.
selection-specification is a filename or in-place string with the selection
specification. Each selection-specification has the following compact notation
format:
}
\details{
\preformatted{ selmulti     Select multiple fields
 delmulti     Delete multiple fields
 changemulti  Change identication of multiple fields
}
}
