% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selvar.R
\name{Selvar}
\alias{Selvar}
\alias{cdo_delcode}
\alias{cdo_delname}
\alias{cdo_delparam}
\alias{cdo_selcode}
\alias{cdo_selgrid}
\alias{cdo_sellevel}
\alias{cdo_sellevidx}
\alias{cdo_selltype}
\alias{cdo_selname}
\alias{cdo_selparam}
\alias{cdo_selstdname}
\alias{cdo_seltabnum}
\alias{cdo_selzaxis}
\alias{cdo_selzaxisname}
\title{Select fields}
\usage{
cdo_delcode(input_file, output_file = NULL)

cdo_delname(input_file, output_file = NULL)

cdo_delparam(input_file, output_file = NULL)

cdo_selcode(input_file, output_file = NULL)

cdo_selgrid(input_file, output_file = NULL)

cdo_sellevel(input_file, output_file = NULL)

cdo_sellevidx(input_file, output_file = NULL)

cdo_selltype(input_file, output_file = NULL)

cdo_selname(input_file, output_file = NULL)

cdo_selparam(input_file, output_file = NULL)

cdo_selstdname(input_file, output_file = NULL)

cdo_seltabnum(input_file, output_file = NULL)

cdo_selzaxis(input_file, output_file = NULL)

cdo_selzaxisname(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module selects some fields from infile and writes them to outfile. The
fields selected depends on the chosen operator and the parameters. A range of
integer values can be specified by first/last\link{/inc}.
}
\details{
\preformatted{ selparam      Select parameters by identifier
               Selects all fields with parameter identifiers in a user given list.
 delparam      Delete parameters by identifier
               Deletes all fields with parameter identifiers in a user given list.
 selcode       Select parameters by code number
               Selects all fields with code numbers in a user given list or range.
 delcode       Delete parameters by code number
               Deletes all fields with code numbers in a user given list or range.
 selname       Select parameters by name
               Selects all fields with parameter names in a user given list.
 delname       Delete parameters by name
               Deletes all fields with parameter names in a user given list.
 selstdname    Select parameters by standard name
               Selects all fields with standard names in a user given list.
 sellevel      Select levels
               Selects all fields with levels in a user given list.
 sellevidx     Select levels by index
               Selects all fields with index of levels in a user given list or range.
 selgrid       Select grids
               Selects all fields with grids in a user given list.
 selzaxis      Select z-axes
               Selects all fields with z-axes in a user given list.
 selzaxisname  Select z-axes by name
               Selects all fields with z-axis names in a user given list.
 selltype      Select GRIB level types
               Selects all fields with GRIB level type in a user given list or range.
 seltabnum     Select parameter table numbers
               Selects all fields with parameter table numbers in a user given list or range.
}
}
