% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sethalo.R
\name{Sethalo}
\alias{Sethalo}
\alias{cdo_sethalo}
\title{Set the left and right bounds of a field}
\usage{
cdo_sethalo(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This operator sets the left and right bounds of the rectangularly understood
fields. Positive numbers of the parameter lhalo enlarges the left bound by the
given number of columns from the right bound. The parameter rhalo does the
similar for the right bound. Negative numbers of the parameter lhalo/rhalo can
be used to remove the given number of columns of the left and right bounds.
}
\details{
\preformatted{ histcount  Histogram count
            Number of elements in the bin range.
 histsum    Histogram sum
            Sum of elements in the bin range.
 histmean   Histogram mean
            Mean of elements in the bin range.
 histfreq   Histogram frequency
            Relative frequency of elements in the bin range.
}
}
