% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftxy.R
\name{Shiftxy}
\alias{Shiftxy}
\alias{cdo_shiftx}
\alias{cdo_shifty}
\title{Shift field}
\usage{
cdo_shiftx(input_file, output_file = NULL)

cdo_shifty(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module contains operators to shift all fields in x or y direction. All
fields need to have the same horizontal rectilinear or curvilinear grid.
}
\details{
\preformatted{ shiftx  Shift x
         Shifts all fields in x direction.
 shifty  Shift y
         Shifts all fields in y direction.
}
}
