% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specconv.R
\name{Specconv}
\alias{Specconv}
\alias{cdo_sp2sp}
\title{Spectral conversion}
\usage{
cdo_sp2sp(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Changed the triangular truncation of all spectral fields. This operator
performs downward conversion by cutting the resolution. Upward conversions are
achieved by filling in zeros.
}
\details{
\preformatted{ sp2gp  Spectral to gridpoint
        Convert all spectral fields to a global regular Gaussian grid.
 gp2sp  Gridpoint to spectral
        Convert all Gaussian gridpoint fields to spectral fields.
}
}
