% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{Spectral}
\alias{Spectral}
\alias{cdo_gp2sp}
\alias{cdo_sp2gp}
\title{Spectral transformation}
\usage{
cdo_gp2sp(infile, outfile = NULL)

cdo_sp2gp(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module transforms fields on a global regular Gaussian grid to spectral
coefficients and vice versa. The transformation is achieved by applying Fast
Fourier Transformation (FFT) first and direct Legendre Transformation
afterwards in gp2sp. In sp2gp the inverse Legendre Transformation and inverse
FFT are used. Missing values are not supported.
}
\details{
\preformatted{ sp2gp  Spectral to gridpoint
        Convert all spectral fields to a global regular Gaussian grid.
 gp2sp  Gridpoint to spectral
        Convert all Gaussian gridpoint fields to spectral fields.
}
}
