% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{Split}
\alias{Split}
\alias{cdo_splitcode}
\alias{cdo_splitgrid}
\alias{cdo_splitlevel}
\alias{cdo_splitname}
\alias{cdo_splitparam}
\alias{cdo_splittabnum}
\alias{cdo_splitzaxis}
\title{Split a dataset}
\usage{
cdo_splitcode(infile, obase = NULL)

cdo_splitgrid(infile, obase = NULL)

cdo_splitlevel(infile, obase = NULL)

cdo_splitname(infile, obase = NULL)

cdo_splitparam(infile, obase = NULL)

cdo_splittabnum(infile, obase = NULL)

cdo_splitzaxis(infile, obase = NULL)
}
\arguments{
\item{infile}{input file.}

\item{obase}{preffix added to the output files.}
}
\description{
This module splits infile into pieces. The output files will be named
\<obase\>\<xxx\>\<suffix\> where suffix is the filename extension derived from
the file format. xxx and the contents of the output files depends on the chosen
operator. params is a comma-separated list of processing parameters.
}
\details{
\preformatted{ splitcode    Split code numbers
              Splits a dataset into pieces, one for each different code number.
              xxx will have three digits with the code number.
 splitparam   Split parameter identifiers
              Splits a dataset into pieces, one for each different parameter identifier.
              xxx will be a string with the parameter identifier.
 splitname    Split variable names
              Splits a dataset into pieces, one for each variable name.
              xxx will be a string with the variable name.
 splitlevel   Split levels
              Splits a dataset into pieces, one for each different level.
              xxx will have six digits with the level.
 splitgrid    Split grids
              Splits a dataset into pieces, one for each different grid.
              xxx will have two digits with the grid number.
 splitzaxis   Split z-axes
              Splits a dataset into pieces, one for each different z-axis.
              xxx will have two digits with the z-axis number.
 splittabnum  Split parameter table numbers
              Splits a dataset into pieces, one for each GRIB1 parameter table number.
              xxx will have three digits with the GRIB1 parameter table number.
}
}
