% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splittime.R
\name{Splittime}
\alias{Splittime}
\alias{cdo_splitday}
\alias{cdo_splithour}
\alias{cdo_splitmon}
\alias{cdo_splitseas}
\alias{cdo_splityear}
\alias{cdo_splityearmon}
\title{Split timesteps of a dataset}
\usage{
cdo_splitday(input_file, output_files = NULL)

cdo_splithour(input_file, output_files = NULL)

cdo_splitmon(input_file, output_files = NULL)

cdo_splitseas(input_file, output_files = NULL)

cdo_splityear(input_file, output_files = NULL)

cdo_splityearmon(input_file, output_files = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_files}{output files.}
}
\description{
This module splits infile into timesteps pieces. The output files will be named
\<obase\>\<xxx\>\<suffix\> where suffix is the filename extension derived from
the file format. xxx and the contents of the output files depends on the chosen
operator.
}
\details{
\preformatted{ splithour     Split hours
               Splits a file into pieces, one for each different hour.
               xxx will have two digits with the hour.
 splitday      Split days
               Splits a file into pieces, one for each different day.
               xxx will have two digits with the day.
 splitseas     Split seasons
               Splits a file into pieces, one for each different season.
               xxx will have three characters with the season.
 splityear     Split years
               Splits a file into pieces, one for each different year.
               xxx will have four digits with the year (YYYY).
 splityearmon  Split in years and months
               Splits a file into pieces, one for each different year and month.
               xxx will have six digits with the year and month (YYYYMM).
 splitmon      Split months
               Splits a file into pieces, one for each different month.
               xxx will have two digits with the month.
}
}
