% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strgal.R
\name{Strgal}
\alias{Strgal}
\alias{cdo_strgal}
\title{Strong gale days index per time period}
\usage{
cdo_strgal(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily maximum horizontal wind speed VX, then
the number of days where VX is greater than or equal to 20.5 m/s is counted. A
further output variable is the maximum number of consecutive days with maximum
wind speed greater than or equal to 20.5 m/s. Note that VX is defined as the
square root of the sum of square of the zonal and meridional wind speeds and
have to be given in units of m/s. The date information of a timestep in outfile
is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ histcount  Histogram count
            Number of elements in the bin range.
 histsum    Histogram sum
            Sum of elements in the bin range.
 histmean   Histogram mean
            Mean of elements in the bin range.
 histfreq   Histogram frequency
            Relative frequency of elements in the bin range.
}
}
