% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwin.R
\name{Strwin}
\alias{Strwin}
\alias{cdo_strwin}
\title{Strong wind days index per time period}
\usage{
cdo_strwin(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
Let infile be a time series of the daily maximum horizontal wind speed VX, then
the number of days where VX \> v is counted. The horizontal wind speed v is an
optional parameter with default v = 10.5 m/s. A further output variable is the
maximum number of consecutive days with maximum wind speed greater than or
equal to v. Note that both VX and v have to be given in units of m/s. Also note
that the horizontal wind speed is defined as the square root of the sum of
squares of the zonal and meridional wind speeds. The date information of a
timestep in outfile is the date of the last contributing timestep in infile.
}
\details{
\preformatted{ histcount  Histogram count
            Number of elements in the bin range.
 histsum    Histogram sum
            Sum of elements in the bin range.
 histmean   Histogram mean
            Mean of elements in the bin range.
 histfreq   Histogram frequency
            Relative frequency of elements in the bin range.
}
}
