% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timcor.R
\name{Timcor}
\alias{Timcor}
\alias{cdo_timcor}
\title{Correlation over time}
\usage{
cdo_timcor(input_file_1, input_file_2, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2}{input files.}

\item{output_file}{output file.}
}
\description{
The correlation coefficient is a quantity that gives the quality of a least
squares fitting to the original data. This operator correlates each gridpoint
of two fields over all timesteps. With
}
\details{
\preformatted{ ydrunmin   Multi-year daily running minimum
            o(001,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x); day[(i(t+(nts-1)/2)] = 001\}
}
}
