% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timcovar.R
\name{Timcovar}
\alias{Timcovar}
\alias{cdo_timcovar}
\title{Covariance over time}
\usage{
cdo_timcovar(infile1, infile2, outfile = NULL)

cdo_timcovar(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This operator calculates the covariance of two fields at each gridpoint over
all timesteps. With
}
\details{
\preformatted{ ydrunmin   Multi-year daily running minimum
            o(001,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x); day[(i(t+(nts-1)/2)] = 001\}
}
}
