% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timcumsum.R
\name{Timcumsum}
\alias{Timcumsum}
\alias{cdo_timcumsum}
\title{Cumulative sum over all timesteps}
\usage{
cdo_timcumsum(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
The timcumsum operator calculates the cumulative sum over all timesteps.
Missing values are treated as numeric zero when summing.
}
\details{
\preformatted{ muldpm  Multiply with days per month
         o(t,x) = i(t,x) * days_per_month
 divdpm  Divide by days per month
         o(t,x) = i(t,x) / days_per_month
 muldpy  Multiply with days per year
         o(t,x) = i(t,x) * days_per_year
 divdpy  Divide by days per year
         o(t,x) = i(t,x) / days_per_year
}
}
