% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timselstat.R
\name{Timselstat}
\alias{Timselstat}
\alias{cdo_timselavg}
\alias{cdo_timselmax}
\alias{cdo_timselmean}
\alias{cdo_timselmin}
\alias{cdo_timselrange}
\alias{cdo_timselstd}
\alias{cdo_timselstd1}
\alias{cdo_timselsum}
\alias{cdo_timselvar}
\alias{cdo_timselvar1}
\title{Time range statistical values}
\usage{
cdo_timselavg(input_file, output_file = NULL)

cdo_timselmax(input_file, output_file = NULL)

cdo_timselmean(input_file, output_file = NULL)

cdo_timselmin(input_file, output_file = NULL)

cdo_timselrange(input_file, output_file = NULL)

cdo_timselstd(input_file, output_file = NULL)

cdo_timselstd1(input_file, output_file = NULL)

cdo_timselsum(input_file, output_file = NULL)

cdo_timselvar(input_file, output_file = NULL)

cdo_timselvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values for a selected number of timesteps.
According to the chosen operator the minimum, maximum, range, sum, average,
variance or standard deviation of the selected timesteps is written to outfile.
The time of outfile is determined by the time in the middle of all contributing
timesteps of infile. This can be change with the CDO option --timestat_date
\<first|middle|last\>.
}
\details{
\preformatted{ timselmin    Time selection minimum
              For every adjacent sequence t1, ...., tn of timesteps of the same selected time range it is:
}
}
