% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstat.R
\name{Timstat}
\alias{Timstat}
\alias{cdo_timavg}
\alias{cdo_timmax}
\alias{cdo_timmean}
\alias{cdo_timmin}
\alias{cdo_timrange}
\alias{cdo_timstd}
\alias{cdo_timstd1}
\alias{cdo_timsum}
\alias{cdo_timvar}
\alias{cdo_timvar1}
\title{Statistical values over all timesteps}
\usage{
cdo_timavg(input_file, output_file = NULL)

cdo_timmax(input_file, output_file = NULL)

cdo_timmean(input_file, output_file = NULL)

cdo_timmin(input_file, output_file = NULL)

cdo_timrange(input_file, output_file = NULL)

cdo_timstd(input_file, output_file = NULL)

cdo_timstd1(input_file, output_file = NULL)

cdo_timsum(input_file, output_file = NULL)

cdo_timvar(input_file, output_file = NULL)

cdo_timvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values over all timesteps in infile. Depending
on the chosen operator the minimum, maximum, range, sum, average, variance or
standard deviation of all timesteps read from infile is written to outfile. The
time of outfile is determined by the time in the middle of all contributing
timesteps of infile. This can be change with the CDO option --timestat_date
\<first|middle|last\>.
}
\details{
\preformatted{ timmin    Time minimum
           o(1,x) = min\{i(t',x), t_1<t'<=t_n\}
 timmax    Time maximum
           o(1,x) = max\{i(t',x), t_1<t'<=t_n\}
 timrange  Time range
           o(1,x) = range\{i(t',x), t_1<t'<=t_n\}
 timsum    Time sum
           o(1,x) = sum\{i(t',x), t_1<t'<=t_n\}
 timmean   Time mean
           o(1,x) = mean\{i(t',x), t_1<t'<=t_n\}
 timavg    Time average
           o(1,x) = avg\{i(t',x), t_1<t'<=t_n\}
 timstd    Time standard deviation
           Normalize by n. 
}
}
