% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{Trend}
\alias{Trend}
\alias{cdo_trend}
\title{Trend of time series}
\usage{
cdo_trend(input_file, output_file_1 = NULL, output_file_2 = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file_1, output_file_2}{output files.}
}
\description{
The values of the input file infile are assumed to be distributed as
N(a+b*t,S^2) with unknown a, b and S^2. This operator estimates the parameter a
and b. For every field element x only those timesteps t belong to the sample
S(x), which have i(t,x) NE miss. Thus the estimation for a is stored in
outfile1 and that for b is stored in outfile2. To subtract the trend from the
data see operator subtrend. It is assumed that all timesteps are equidistant,
if this is not the case set the parameter equal=false.
}
\details{
\preformatted{ ydrunmin   Multi-year daily running minimum
            o(001,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x); day[(i(t+(nts-1)/2)] = 001\}
}
}
