% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertintml.R
\name{Vertintml}
\alias{Vertintml}
\alias{cdo_ml2hl}
\alias{cdo_ml2pl}
\title{Vertical interpolation}
\usage{
cdo_ml2hl(infile, outfile = NULL)

cdo_ml2pl(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Interpolate 3D variables on hybrid sigma pressure level to pressure or height
levels. The input file should contain the log. surface pressure or the surface
pressure. To extrapolate the temperature, the surface geopotential is also
needed. The pressure, temperature, and surface geopotential are identified by
their GRIB1 code number or NetCDF CF standard name. Supported parameter tables
are: WMO standard table number 2 and ECMWF local table number 128. Use the
alias ml2plx/ml2hlx or the environment variable EXTRAPOLATE to extrapolate
missing values. This operator requires all variables on the same horizontal
grid.
}
\details{
\preformatted{ ml2pl  Model to pressure level interpolation
        Interpolates 3D variables on hybrid sigma pressure level to pressure level.
 ml2hl  Model to height level interpolation
        Interpolates 3D variables on hybrid sigma pressure level to height level.
        The procedure is the same as for the operator ml2pl except for
        the pressure levels being calculated from the heights by:
        plevel = 101325*exp(hlevel/-7000)
}
}
