% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind2.R
\name{Wind2}
\alias{Wind2}
\alias{cdo_dv2ps}
\title{D and V to velocity potential and stream function}
\usage{
cdo_dv2ps(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
Calculate spherical harmonic coefficients of velocity potential and stream
function from spherical harmonic coefficients of relative divergence and
vorticity. The divergence and vorticity need to have the names sd and svo or
code numbers 155 and 138.
}
\details{
\preformatted{ sp2gp  Spectral to gridpoint
        Convert all spectral fields to a global regular Gaussian grid.
 gp2sp  Gridpoint to spectral
        Convert all Gaussian gridpoint fields to spectral fields.
}
}
