% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydaystat.R
\name{Ydaystat}
\alias{Ydaystat}
\alias{cdo_ydayavg}
\alias{cdo_ydaymax}
\alias{cdo_ydaymean}
\alias{cdo_ydaymin}
\alias{cdo_ydayrange}
\alias{cdo_ydaystd}
\alias{cdo_ydaystd1}
\alias{cdo_ydaysum}
\alias{cdo_ydayvar}
\alias{cdo_ydayvar1}
\title{Multi-year daily statistical values}
\usage{
cdo_ydayavg(infile, outfile = NULL)

cdo_ydaymax(infile, outfile = NULL)

cdo_ydaymean(infile, outfile = NULL)

cdo_ydaymin(infile, outfile = NULL)

cdo_ydayrange(infile, outfile = NULL)

cdo_ydaystd(infile, outfile = NULL)

cdo_ydaystd1(infile, outfile = NULL)

cdo_ydaysum(infile, outfile = NULL)

cdo_ydayvar(infile, outfile = NULL)

cdo_ydayvar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes statistical values of each day of year. Depending on the
chosen operator the minimum, maximum, range, sum, average, variance or standard
deviation of each day of year in infile is written to outfile. The date
information in an output field is the date of the last contributing input
field.
}
\details{
\preformatted{ ydaymin    Multi-year daily minimum
            o(001,x) = min\{i(t,x), day(i(t)) = 001\}
}
}
