% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearstat.R
\name{Yearstat}
\alias{Yearstat}
\alias{cdo_yearavg}
\alias{cdo_yearmax}
\alias{cdo_yearmean}
\alias{cdo_yearmin}
\alias{cdo_yearrange}
\alias{cdo_yearstd}
\alias{cdo_yearstd1}
\alias{cdo_yearsum}
\alias{cdo_yearvar}
\alias{cdo_yearvar1}
\title{Yearly statistical values}
\usage{
cdo_yearavg(input_file, output_file = NULL)

cdo_yearmax(input_file, output_file = NULL)

cdo_yearmean(input_file, output_file = NULL)

cdo_yearmin(input_file, output_file = NULL)

cdo_yearrange(input_file, output_file = NULL)

cdo_yearstd(input_file, output_file = NULL)

cdo_yearstd1(input_file, output_file = NULL)

cdo_yearsum(input_file, output_file = NULL)

cdo_yearvar(input_file, output_file = NULL)

cdo_yearvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values over timesteps of the same year.
Depending on the chosen operator the minimum, maximum, range, sum, average,
variance or standard deviation of timesteps of the same year is written to
outfile. The time of outfile is determined by the time in the middle of all
contributing timesteps of infile. This can be change with the CDO option
--timestat_date \<first|middle|last\>.
}
\details{
\preformatted{ yearmin     Yearly minimum
             For every adjacent sequence t_1, ...,t_n of timesteps of the same year it is:
}
}
