% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhourarith.R
\name{Yhourarith}
\alias{Yhourarith}
\alias{cdo_yhouradd}
\alias{cdo_yhourdiv}
\alias{cdo_yhourmul}
\alias{cdo_yhoursub}
\title{Multi-year hourly arithmetic}
\usage{
cdo_yhouradd(infile1, infile2, outfile = NULL)

cdo_yhouradd(infile1, infile2, outfile = NULL)

cdo_yhourdiv(infile1, infile2, outfile = NULL)

cdo_yhourdiv(infile1, infile2, outfile = NULL)

cdo_yhourmul(infile1, infile2, outfile = NULL)

cdo_yhourmul(infile1, infile2, outfile = NULL)

cdo_yhoursub(infile1, infile2, outfile = NULL)

cdo_yhoursub(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This module performs simple arithmetic of a time series and one timestep with
the same hour and day of year. For each field in infile1 the corresponding
field of the timestep in infile2 with the same hour and day of year is used.
The header information in infile1 have to be the same as in infile2. Usually
infile2 is generated by an operator of the module YHOURSTAT.
}
\details{
\preformatted{ yhouradd  Add multi-year hourly time series
           Adds a time series and a multi-year hourly time series.
 yhoursub  Subtract multi-year hourly time series
           Subtracts a time series and a multi-year hourly time series.
 yhourmul  Multiply multi-year hourly time series
           Multiplies a time series and a multi-year hourly time series.
 yhourdiv  Divide multi-year hourly time series
           Divides a time series and a multi-year hourly time series.
}
}
