% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhourstat.R
\name{Yhourstat}
\alias{Yhourstat}
\alias{cdo_yhouravg}
\alias{cdo_yhourmax}
\alias{cdo_yhourmean}
\alias{cdo_yhourmin}
\alias{cdo_yhourrange}
\alias{cdo_yhourstd}
\alias{cdo_yhourstd1}
\alias{cdo_yhoursum}
\alias{cdo_yhourvar}
\alias{cdo_yhourvar1}
\title{Multi-year hourly statistical values}
\usage{
cdo_yhouravg(input_file, output_file = NULL)

cdo_yhourmax(input_file, output_file = NULL)

cdo_yhourmean(input_file, output_file = NULL)

cdo_yhourmin(input_file, output_file = NULL)

cdo_yhourrange(input_file, output_file = NULL)

cdo_yhourstd(input_file, output_file = NULL)

cdo_yhourstd1(input_file, output_file = NULL)

cdo_yhoursum(input_file, output_file = NULL)

cdo_yhourvar(input_file, output_file = NULL)

cdo_yhourvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values of each hour and day of year. Depending
on the chosen operator the minimum, maximum, range, sum, average, variance or
standard deviation of each hour and day of year in infile is written to
outfile. The date information in an output field is the date of the last
contributing input field.
}
\details{
\preformatted{ yhourmin    Multi-year hourly minimum
             o(0001,x) = min\{i(t,x), day(i(t)) = 0001\}
}
}
