% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yseasarith.R
\name{Yseasarith}
\alias{Yseasarith}
\alias{cdo_yseasadd}
\alias{cdo_yseasdiv}
\alias{cdo_yseasmul}
\alias{cdo_yseassub}
\title{Multi-year seasonal arithmetic}
\usage{
cdo_yseasadd(infile1, infile2, outfile = NULL)

cdo_yseasadd(infile1, infile2, outfile = NULL)

cdo_yseasdiv(infile1, infile2, outfile = NULL)

cdo_yseasdiv(infile1, infile2, outfile = NULL)

cdo_yseasmul(infile1, infile2, outfile = NULL)

cdo_yseasmul(infile1, infile2, outfile = NULL)

cdo_yseassub(infile1, infile2, outfile = NULL)

cdo_yseassub(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This module performs simple arithmetic of a time series and one timestep with
the same season. For each field in infile1 the corresponding field of the
timestep in infile2 with the same season is used. The header information in
infile1 have to be the same as in infile2. Usually infile2 is generated by an
operator of the module YSEASSTAT.
}
\details{
\preformatted{ yseasadd  Add multi-year seasonal time series
           Adds a time series and a multi-year seasonal time series.
 yseassub  Subtract multi-year seasonal time series
           Subtracts a time series and a multi-year seasonal time series.
 yseasmul  Multiply multi-year seasonal time series
           Multiplies a time series and a multi-year seasonal time series.
 yseasdiv  Divide multi-year seasonal time series
           Divides a time series and a multi-year seasonal time series.
}
}
