% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldcor.R
\name{Fldcor}
\alias{Fldcor}
\alias{cdo_fldcor}
\title{Correlation in grid space}
\usage{
cdo_fldcor(infile1, infile2, outfile = NULL)

cdo_fldcor(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
The correlation coefficient is a quantity that gives the quality of a least
squares fitting to the original data. This operator correlates all gridpoints
of two fields for each timestep. With
}
\details{
\preformatted{ ydrunmin   Multi-year daily running minimum
            o(001,x) = min\{i(t,x), i(t+1,x), ..., i(t+nts-1,x); day[(i(t+(nts-1)/2)] = 001\}
}
}
