% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridboxstat.R
\name{Gridboxstat}
\alias{Gridboxstat}
\alias{cdo_gridboxavg}
\alias{cdo_gridboxmax}
\alias{cdo_gridboxmean}
\alias{cdo_gridboxmin}
\alias{cdo_gridboxrange}
\alias{cdo_gridboxstd}
\alias{cdo_gridboxstd1}
\alias{cdo_gridboxsum}
\alias{cdo_gridboxvar}
\alias{cdo_gridboxvar1}
\title{Statistical values over grid boxes}
\usage{
cdo_gridboxavg(input_file, output_file = NULL)

cdo_gridboxmax(input_file, output_file = NULL)

cdo_gridboxmean(input_file, output_file = NULL)

cdo_gridboxmin(input_file, output_file = NULL)

cdo_gridboxrange(input_file, output_file = NULL)

cdo_gridboxstd(input_file, output_file = NULL)

cdo_gridboxstd1(input_file, output_file = NULL)

cdo_gridboxsum(input_file, output_file = NULL)

cdo_gridboxvar(input_file, output_file = NULL)

cdo_gridboxvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values over surrounding grid boxes. According
to the chosen operator the minimum, maximum, range, sum, average, variance, or
standard deviation of the neighboring grid boxes is written to outfile. All
gridbox operators only works on quadrilateral curvilinear grids.
}
\details{
\preformatted{ gridboxmin    Gridbox minimum
               Minimum value of the selected grid boxes.
 gridboxmax    Gridbox maximum
               Maximum value of the selected grid boxes.
 gridboxrange  Gridbox range
               Range (max-min value) of the selected grid boxes.
 gridboxsum    Gridbox sum
               Sum of the selected grid boxes.
 gridboxmean   Gridbox mean
               Mean of the selected grid boxes.
 gridboxavg    Gridbox average
               Average of the selected grid boxes.
 gridboxstd    Gridbox standard deviation
               Standard deviation of the selected grid boxes. Normalize by n.
 gridboxstd1   Gridbox standard deviation (n-1)
               Standard deviation of the selected grid boxes. Normalize by (n-1).
 gridboxvar    Gridbox variance
               Variance of the selected grid boxes. Normalize by n.
 gridboxvar1   Gridbox variance (n-1)
               Variance of the selected grid boxes. Normalize by (n-1).
}
}
