% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inttime.R
\name{Inttime}
\alias{Inttime}
\alias{cdo_intntime}
\alias{cdo_inttime}
\title{Time interpolation}
\usage{
cdo_intntime(infile, outfile = NULL)

cdo_inttime(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module performs linear interpolation between timesteps.
}
\details{
\preformatted{ inttime   Interpolation between timesteps
           This operator creates a new dataset by linear interpolation between timesteps.
           The user has to define the start date/time with an optional increment.
 intntime  Interpolation between timesteps
           This operator performs linear interpolation between timesteps.
           The user has to define the number of timesteps from one timestep to the next.
}
}
