% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{Math}
\alias{Math}
\alias{cdo_abs}
\alias{cdo_acos}
\alias{cdo_asin}
\alias{cdo_atan}
\alias{cdo_cos}
\alias{cdo_exp}
\alias{cdo_int}
\alias{cdo_ln}
\alias{cdo_log10}
\alias{cdo_nint}
\alias{cdo_not}
\alias{cdo_pow}
\alias{cdo_reci}
\alias{cdo_sin}
\alias{cdo_sqr}
\alias{cdo_sqrt}
\alias{cdo_tan}
\title{Mathematical functions}
\usage{
cdo_abs(infile, outfile = NULL)

cdo_acos(infile, outfile = NULL)

cdo_asin(infile, outfile = NULL)

cdo_atan(infile, outfile = NULL)

cdo_cos(infile, outfile = NULL)

cdo_exp(infile, outfile = NULL)

cdo_int(infile, outfile = NULL)

cdo_ln(infile, outfile = NULL)

cdo_log10(infile, outfile = NULL)

cdo_nint(infile, outfile = NULL)

cdo_not(infile, outfile = NULL)

cdo_pow(infile, outfile = NULL)

cdo_reci(infile, outfile = NULL)

cdo_sin(infile, outfile = NULL)

cdo_sqr(infile, outfile = NULL)

cdo_sqrt(infile, outfile = NULL)

cdo_tan(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module contains some standard mathematical functions. All trigonometric
functions calculate with radians.
}
\details{
\preformatted{ abs    Absolute value
        o(t,x) = abs(i(t,x))
 int    Integer value
        o(t,x) = int(i(t,x))
 nint   Nearest integer value
        o(t,x) = nint(i(t,x))
 pow    Power
        o(t,x) = i(t,x)^y
 sqr    Square
        o(t,x) = i(t,x)^2
 sqrt   Square root
        o(t,x) = sqrt(i(t,x))
 exp    Exponential
        o(t,x) = e^i(t,x)
 ln     Natural logarithm
        o(t,x) = ln(i(t,x))
 log10  Base 10 logarithm
        o(t,x) = log10(i(t,x))
 sin    Sine
        o(t,x) = sin(i(t,x))
 cos    Cosine
        o(t,x) = cos(i(t,x))
 tan    Tangent
        o(t,x) = tan(i(t,x))
 asin   Arc sine
        o(t,x) = asin(i(t,x))
 acos   Arc cosine
        o(t,x) = acos(i(t,x))
 atan   Arc tangent
        o(t,x) = atan(i(t,x))
 reci   Reciprocal value
        o(t,x) = 1 / i(t,x)
 not    Logical NOT
        o(t,x) = 1, if x equal 0; else 0
}
}
