% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monstat.R
\name{Monstat}
\alias{Monstat}
\alias{cdo_monavg}
\alias{cdo_monmax}
\alias{cdo_monmean}
\alias{cdo_monmin}
\alias{cdo_monrange}
\alias{cdo_monstd}
\alias{cdo_monstd1}
\alias{cdo_monsum}
\alias{cdo_monvar}
\alias{cdo_monvar1}
\title{Monthly statistical values}
\usage{
cdo_monavg(infile, outfile = NULL)

cdo_monmax(infile, outfile = NULL)

cdo_monmean(infile, outfile = NULL)

cdo_monmin(infile, outfile = NULL)

cdo_monrange(infile, outfile = NULL)

cdo_monstd(infile, outfile = NULL)

cdo_monstd1(infile, outfile = NULL)

cdo_monsum(infile, outfile = NULL)

cdo_monvar(infile, outfile = NULL)

cdo_monvar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes statistical values over timesteps of the same month.
Depending on the chosen operator the minimum, maximum, range, sum, average,
variance or standard deviation of timesteps of the same month is written to
outfile. The time of outfile is determined by the time in the middle of all
contributing timesteps of infile. This can be change with the CDO option
--timestat_date \<first|middle|last\>.
}
\details{
\preformatted{ monmin    Monthly minimum
           For every adjacent sequence t_1, ...,t_n of timesteps of the same month it is:
}
}
