% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninfo.R
\name{Ninfo}
\alias{Ninfo}
\alias{cdo_ndate}
\alias{cdo_ngridpoints}
\alias{cdo_ngrids}
\alias{cdo_nlevel}
\alias{cdo_nmon}
\alias{cdo_npar}
\alias{cdo_ntime}
\alias{cdo_nyear}
\title{Print the number of parameters, levels or times}
\usage{
cdo_ndate(input_file)

cdo_ngridpoints(input_file)

cdo_ngrids(input_file)

cdo_nlevel(input_file)

cdo_nmon(input_file)

cdo_npar(input_file)

cdo_ntime(input_file)

cdo_nyear(input_file)
}
\arguments{
\item{input_file}{input file.}
}
\description{
This module prints the number of variables, levels or times of the input
dataset.
}
\details{
\preformatted{ npar         Number of parameters
              Prints the number of parameters (variables).
 nlevel       Number of levels
              Prints the number of levels for each variable.
 nyear        Number of years
              Prints the number of different years.
 nmon         Number of months
              Prints the number of different combinations of years and months.
 ndate        Number of dates
              Prints the number of different dates.
 ntime        Number of timesteps
              Prints the number of timesteps.
 ngridpoints  Number of gridpoints
              Prints the number of gridpoints for each variable.
 ngrids       Number of horizontal grids
              Prints the number of horizontal grids.
}
}
