% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapnn.R
\name{Remapnn}
\alias{Remapnn}
\alias{cdo_gennn}
\alias{cdo_remapnn}
\title{Nearest neighbor remapping}
\usage{
cdo_gennn(input_file, output_file = NULL)

cdo_remapnn(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module contains operators for a nearest neighbor remapping of fields
between grids in spherical coordinates.
}
\details{
\preformatted{ remapnn  Nearest neighbor remapping
          Performs a nearest neighbor remapping on all input fields.
 gennn    Generate nearest neighbor remap weights
          Generates nearest neighbor remapping weights for the first input field and writes the result to a file.
          The format of this file is NetCDF following the SCRIP convention.
          Use the operator remap to apply this remapping weights to a data file with the same source grid.
}
}
