% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{Replace}
\alias{Replace}
\alias{cdo_replace}
\title{Replace variables}
\usage{
cdo_replace(infile1, infile2, outfile = NULL)

cdo_replace(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
The replace operator replaces variables in infile1 by variables from infile2
and write the result to outfile. Both input datasets need to have the same
number of timesteps.
}
\details{
\preformatted{ copy  Copy datasets
       Copies all input datasets to outfile. 
 cat   Concatenate datasets
       Concatenates all input datasets and appends the result to the end 
       of outfile. If outfile does not exist it will be created.
}
}
