% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{Set}
\alias{Set}
\alias{cdo_setcode}
\alias{cdo_setcodetab}
\alias{cdo_setlevel}
\alias{cdo_setltype}
\alias{cdo_setname}
\alias{cdo_setparam}
\alias{cdo_setunit}
\title{Set field info}
\usage{
cdo_setcode(infile, outfile = NULL)

cdo_setcodetab(infile, outfile = NULL)

cdo_setlevel(infile, outfile = NULL)

cdo_setltype(infile, outfile = NULL)

cdo_setname(infile, outfile = NULL)

cdo_setparam(infile, outfile = NULL)

cdo_setunit(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module sets some field information. Depending on the chosen operator the
parameter table, code number, parameter identifier, variable name or level is
set.
}
\details{
\preformatted{ setcodetab  Set parameter code table
             Sets the parameter code table for all variables.
 setcode     Set code number
             Sets the code number for all variables to the same given value.
 setparam    Set parameter identifier
             Sets the parameter identifier of the first variable.
 setname     Set variable name
             Sets the name of the first variable.
 setunit     Set variable unit
             Sets the unit of the first variable.
 setlevel    Set level
             Sets the first level of all variables.
 setltype    Set GRIB level type
             Sets the GRIB level type of all variables.
}
}
