% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpartab.R
\name{Setpartab}
\alias{Setpartab}
\alias{cdo_setpartabn}
\alias{cdo_setpartabp}
\title{Set parameter table}
\usage{
cdo_setpartabn(input_file, output_file = NULL)

cdo_setpartabp(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module transforms data and metadata of infile via a parameter table and
writes the result to outfile. A parameter table is an ASCII formatted file with
a set of parameter entries for each variable. Each new set have to start with
\"\\&parameter\" and to end with \"/\".
}
\details{
\preformatted{ setpartabp  Set parameter table
             Search variables by the parameter identifier.
 setpartabn  Set parameter table
             Search variables by name.
}
}
