% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setzaxis.R
\name{Setzaxis}
\alias{Setzaxis}
\alias{cdo_genlevelbounds}
\alias{cdo_setzaxis}
\title{Set z-axis information}
\usage{
cdo_genlevelbounds(input_file, output_file = NULL)

cdo_setzaxis(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module modifies the metadata of the vertical grid.
}
\details{
\preformatted{ setzaxis        Set z-axis
                 This operator sets the z-axis description of all variables with the same number of level as the new z-axis.
 genlevelbounds  Generate level bounds
                 Generates the layer bounds of the z-axis.
}
}
