% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showinfo.R
\name{Showinfo}
\alias{Showinfo}
\alias{cdo_showatts}
\alias{cdo_showattsglob}
\alias{cdo_showcode}
\alias{cdo_showdate}
\alias{cdo_showformat}
\alias{cdo_showlevel}
\alias{cdo_showltype}
\alias{cdo_showmon}
\alias{cdo_showname}
\alias{cdo_showstdname}
\alias{cdo_showtime}
\alias{cdo_showtimestamp}
\alias{cdo_showyear}
\title{Show variables, levels or times}
\usage{
cdo_showatts(infile)

cdo_showattsglob(infile)

cdo_showcode(infile)

cdo_showdate(infile)

cdo_showformat(infile)

cdo_showlevel(infile)

cdo_showltype(infile)

cdo_showmon(infile)

cdo_showname(infile)

cdo_showstdname(infile)

cdo_showtime(infile)

cdo_showtimestamp(infile)

cdo_showyear(infile)
}
\arguments{
\item{infile}{input file.}
}
\description{
This module prints the format, variables, levels or times of the input dataset.
}
\details{
\preformatted{ showformat     Show file format
                Prints the file format of the input dataset.
 showcode       Show code numbers
                Prints the code number of all variables.
 showname       Show variable names
                Prints the name of all variables.
 showstdname    Show standard names
                Prints the standard name of all variables.
 showatts       Show all attributes
                Prints all variable and global attributes.
 showattsglob   Show all global attributes
                Prints all global attributes.
 showlevel      Show levels
                Prints all levels for each variable.
 showltype      Show GRIB level types
                Prints the GRIB level type for all z-axes.
 showyear       Show years
                Prints all years.
 showmon        Show months
                Prints all months.
 showdate       Show date information
                Prints date information of all timesteps (format YYYY-MM-DD).
 showtime       Show time information
                Prints time information of all timesteps (format hh:mm:ss).
 showtimestamp  Show timestamp
                Prints timestamp of all timesteps (format YYYY-MM-DDThh:mm:ss).
}
}
