% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tee.R
\name{Tee}
\alias{Tee}
\alias{cdo_tee}
\title{Duplicate a data stream and write it to file}
\usage{
cdo_tee(infile1, infile2, outfile = NULL)

cdo_tee(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This operator copies the input dataset to outfile1 and outfile2. The first
output stream in outfile1 can be further processesd with other cdo operators.
The second output outfile2 is written to disk. It can be used to store
intermediate results to a file.
}
\details{
\preformatted{ copy  Copy datasets
       Copies all input datasets to outfile. 
 cat   Concatenate datasets
       Concatenates all input datasets and appends the result to the end 
       of outfile. If outfile does not exist it will be created.
}
}
