% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertstat.R
\name{Vertstat}
\alias{Vertstat}
\alias{cdo_vertavg}
\alias{cdo_vertmax}
\alias{cdo_vertmean}
\alias{cdo_vertmin}
\alias{cdo_vertrange}
\alias{cdo_vertstd}
\alias{cdo_vertstd1}
\alias{cdo_vertsum}
\alias{cdo_vertvar}
\alias{cdo_vertvar1}
\title{Vertical statistical values}
\usage{
cdo_vertavg(infile, outfile = NULL)

cdo_vertmax(infile, outfile = NULL)

cdo_vertmean(infile, outfile = NULL)

cdo_vertmin(infile, outfile = NULL)

cdo_vertrange(infile, outfile = NULL)

cdo_vertstd(infile, outfile = NULL)

cdo_vertstd1(infile, outfile = NULL)

cdo_vertsum(infile, outfile = NULL)

cdo_vertvar(infile, outfile = NULL)

cdo_vertvar1(infile, outfile = NULL)
}
\arguments{
\item{infile}{input file.}

\item{outfile}{output file.}
}
\description{
This module computes statistical values over all levels of the input variables.
According to chosen operator the vertical minimum, maximum, range, sum,
average, variance or standard deviation is written to outfile.
}
\details{
\preformatted{ vertmin    Vertical minimum
            For every gridpoint the minimum over all levels is computed.
 vertmax    Vertical maximum
            For every gridpoint the maximum over all levels is computed.
 vertrange  Vertical range
            For every gridpoint the range over all levels is computed.
 vertsum    Vertical sum
            For every gridpoint the sum over all levels is computed.
 vertmean   Vertical mean
            For every gridpoint the layer weighted mean over all levels is computed.
 vertavg    Vertical average
            For every gridpoint the layer weighted average over all levels is computed.
 vertstd    Vertical standard deviation
            For every gridpoint the standard deviation over all levels is computed. Normalize by n.
 vertstd1   Vertical standard deviation (n-1)
            For every gridpoint the standard deviation over all levels is computed. Normalize by (n-1).
 vertvar    Vertical variance
            For every gridpoint the variance over all levels is computed. Normalize by n.
 vertvar1   Vertical variance (n-1)
            For every gridpoint the variance over all levels is computed. Normalize by (n-1).
}
}
