% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wct.R
\name{Wct}
\alias{Wct}
\alias{cdo_wct}
\title{Windchill temperature}
\usage{
cdo_wct(infile1, infile2, outfile = NULL)

cdo_wct(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
Let infile1 and infile2 be time series of temperature and wind speed records,
then a corresponding time series of resulting windchill temperatures is written
to outfile. The wind chill temperature calculation is only valid for a
temperature of T \<= 33 °C and a wind speed of v \>= 1.39 m/s. Whenever these
conditions are not satisfied, a missing value is written to outfile. Note that
temperature and wind speed records have to be given in units of °C and m/s,
respectively.
}
\details{
\preformatted{ histcount  Histogram count
            Number of elements in the bin range.
 histsum    Histogram sum
            Sum of elements in the bin range.
 histmean   Histogram mean
            Mean of elements in the bin range.
 histfreq   Histogram frequency
            Relative frequency of elements in the bin range.
}
}
