% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydayarith.R
\name{Ydayarith}
\alias{Ydayarith}
\alias{cdo_ydayadd}
\alias{cdo_ydaydiv}
\alias{cdo_ydaymul}
\alias{cdo_ydaysub}
\title{Multi-year daily arithmetic}
\usage{
cdo_ydayadd(infile1, infile2, outfile = NULL)

cdo_ydayadd(infile1, infile2, outfile = NULL)

cdo_ydaydiv(infile1, infile2, outfile = NULL)

cdo_ydaydiv(infile1, infile2, outfile = NULL)

cdo_ydaymul(infile1, infile2, outfile = NULL)

cdo_ydaymul(infile1, infile2, outfile = NULL)

cdo_ydaysub(infile1, infile2, outfile = NULL)

cdo_ydaysub(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This module performs simple arithmetic of a time series and one timestep with
the same day of year. For each field in infile1 the corresponding field of the
timestep in infile2 with the same day of year is used. The header information
in infile1 have to be the same as in infile2. Usually infile2 is generated by
an operator of the module YDAYSTAT.
}
\details{
\preformatted{ ydayadd  Add multi-year daily time series
          Adds a time series and a multi-year daily time series.
 ydaysub  Subtract multi-year daily time series
          Subtracts a time series and a multi-year daily time series.
 ydaymul  Multiply multi-year daily time series
          Multiplies a time series and a multi-year daily time series.
 ydaydiv  Divide multi-year daily time series
          Divides a time series and a multi-year daily time series.
}
}
