% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydaypctl.R
\name{Ydaypctl}
\alias{Ydaypctl}
\alias{cdo_ydaypctl}
\title{Multi-year daily percentile values}
\usage{
cdo_ydaypctl(input_file_1, input_file_2, input_file_3, output_file = NULL)
}
\arguments{
\item{input_file_1, input_file_2, input_file_3}{input files.}

\item{output_file}{output file.}
}
\description{
This operator writes a certain percentile of each day of year in infile1 to
outfile. The algorithm uses histograms with minimum and maximum bounds given in
infile2 and infile3, respectively. The default number of histogram bins is 101.
The default can be overridden by setting the environment variable
CDO_PCTL_NBINS to a different value. The files infile2 and infile3 should be
the result of corresponding ydaymin and ydaymax operations, respectively. The
date information in an output field is the date of the last contributing input
field.
}
\details{
\preformatted{ ydaymin    Multi-year daily minimum
            o(001,x) = min\{i(t,x), day(i(t)) = 001\}
}
}
