% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymonarith.R
\name{Ymonarith}
\alias{Ymonarith}
\alias{cdo_ymonadd}
\alias{cdo_ymondiv}
\alias{cdo_ymonmul}
\alias{cdo_ymonsub}
\title{Multi-year monthly arithmetic}
\usage{
cdo_ymonadd(infile1, infile2, outfile = NULL)

cdo_ymonadd(infile1, infile2, outfile = NULL)

cdo_ymondiv(infile1, infile2, outfile = NULL)

cdo_ymondiv(infile1, infile2, outfile = NULL)

cdo_ymonmul(infile1, infile2, outfile = NULL)

cdo_ymonmul(infile1, infile2, outfile = NULL)

cdo_ymonsub(infile1, infile2, outfile = NULL)

cdo_ymonsub(infile1, infile2, outfile = NULL)
}
\arguments{
\item{infile1, infile2}{input files.}

\item{outfile}{output file.}
}
\description{
This module performs simple arithmetic of a time series and one timestep with
the same month of year. For each field in infile1 the corresponding field of
the timestep in infile2 with the same month of year is used. The header
information in infile1 have to be the same as in infile2. Usually infile2 is
generated by an operator of the module YMONSTAT.
}
\details{
\preformatted{ ymonadd  Add multi-year monthly time series
          Adds a time series and a multi-year monthly time series.
 ymonsub  Subtract multi-year monthly time series
          Subtracts a time series and a multi-year monthly time series.
 ymonmul  Multiply multi-year monthly time series
          Multiplies a time series and a multi-year monthly time series.
 ymondiv  Divide multi-year monthly time series
          Divides a time series and a multi-year monthly time series.
}
}
