% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymonstat.R
\name{Ymonstat}
\alias{Ymonstat}
\alias{cdo_ymonavg}
\alias{cdo_ymonmax}
\alias{cdo_ymonmean}
\alias{cdo_ymonmin}
\alias{cdo_ymonrange}
\alias{cdo_ymonstd}
\alias{cdo_ymonstd1}
\alias{cdo_ymonsum}
\alias{cdo_ymonvar}
\alias{cdo_ymonvar1}
\title{Multi-year monthly statistical values}
\usage{
cdo_ymonavg(input_file, output_file = NULL)

cdo_ymonmax(input_file, output_file = NULL)

cdo_ymonmean(input_file, output_file = NULL)

cdo_ymonmin(input_file, output_file = NULL)

cdo_ymonrange(input_file, output_file = NULL)

cdo_ymonstd(input_file, output_file = NULL)

cdo_ymonstd1(input_file, output_file = NULL)

cdo_ymonsum(input_file, output_file = NULL)

cdo_ymonvar(input_file, output_file = NULL)

cdo_ymonvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes statistical values of each month of year. Depending on the
chosen operator the minimum, maximum, range, sum, average, variance or standard
deviation of each month of year in infile is written to outfile. The date
information in an output field is the date of the last contributing input
field.
}
\details{
\preformatted{ ymonmin    Multi-year monthly minimum
            o(01,x) = min\{i(t,x), month(i(t)) = 01\}
}
}
