% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonstat.R
\name{Zonstat}
\alias{Zonstat}
\alias{cdo_zonavg}
\alias{cdo_zonmax}
\alias{cdo_zonmean}
\alias{cdo_zonmin}
\alias{cdo_zonpctl}
\alias{cdo_zonrange}
\alias{cdo_zonstd}
\alias{cdo_zonstd1}
\alias{cdo_zonsum}
\alias{cdo_zonvar}
\alias{cdo_zonvar1}
\title{Zonal statistical values}
\usage{
cdo_zonavg(input_file, output_file = NULL)

cdo_zonmax(input_file, output_file = NULL)

cdo_zonmean(input_file, output_file = NULL)

cdo_zonmin(input_file, output_file = NULL)

cdo_zonpctl(input_file, output_file = NULL)

cdo_zonrange(input_file, output_file = NULL)

cdo_zonstd(input_file, output_file = NULL)

cdo_zonstd1(input_file, output_file = NULL)

cdo_zonsum(input_file, output_file = NULL)

cdo_zonvar(input_file, output_file = NULL)

cdo_zonvar1(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{input file.}

\item{output_file}{output file.}
}
\description{
This module computes zonal statistical values of the input fields. According to
the chosen operator the zonal minimum, maximum, range, sum, average, variance,
standard deviation or a certain percentile is written to outfile. This operator
requires all variables on the same regular lon/lat grid.
}
\details{
\preformatted{ zonmin    Zonal minimum
           For every latitude the minimum over all longitudes is computed.
 zonmax    Zonal maximum
           For every latitude the maximum over all longitudes is computed.
 zonrange  Zonal range
           For every latitude the range over all longitudes is computed.
 zonsum    Zonal sum
           For every latitude the sum over all longitudes is computed.
 zonmean   Zonal mean
           For every latitude the mean over all longitudes is computed.
 zonavg    Zonal average
           For every latitude the average over all longitudes is computed.
 zonstd    Zonal standard deviation
           For every latitude the standard deviation over all longitudes is computed. Normalize by n.
 zonstd1   Zonal standard deviation (n-1)
           For every latitude the standard deviation over all longitudes is computed. Normalize by (n-1). 
 zonvar    Zonal variance
           For every latitude the variance over all longitudes is computed. Normalize by n.
 zonvar1   Zonal variance (n-1)
           For every latitude the variance over all longitudes is computed. Normalize by (n-1).
 zonpctl   Zonal percentiles
           For every latitude the pth percentile over all longitudes is computed.
}
}
